% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{adduct<-}
\alias{adduct<-}
\alias{adduct<-,MSP,character-method}
\title{adduct<- sets adduct ion names in MSP object}
\format{An object of class \code{NULL} of length 0.}
\arguments{
\item{x}{object of class MSP, see ?convert2MSP for further information}

\item{value}{character vector with new adduct ion names}
}
\value{
MSP
}
\description{
adduct<- sets adduct ion names in MSP object
}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", 
                         splitIndMZ = 2, splitIndRT = NULL)
adduct(finalMSP) <- rep("Unknown")
}
\keyword{datasets}

