% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocatePrecursor2mz.R
\name{allocatePrecursor2mz}
\alias{allocatePrecursor2mz}
\title{allocatePrecursor2mz: Join two data sources}
\usage{
allocatePrecursor2mz(sd01, sd02, kNN = 10, mzCheck = 1, rtCheck = 30, mzVsRTbalance = 10000)
}
\arguments{
\item{sd01}{has}

\item{sd02}{has four columns, the first column contains the m/z 
value, the second column the rt, the third column the intensity, the fourth
column the pcgroup_precursorMZ}

\item{kNN}{numerical, number of k-nearest neighbours based on deviation
from m/z (i.e. the kNN entries with the smallest deviation)}

\item{mzCheck}{numerical, maximum tolerated distance for m/z (strong criterion here)}

\item{rtCheck}{numerical, maximum tolerated distance for retention time}

\item{mzVsRTbalance}{numerical, multiplicator for mz value before calculating 
the (euclidean) distance between two peaks, high value means that there is 
a strong weight on the deviation m/z value}
}
\description{
Allocates precursor ions to candidate m / z values based on 
minimal distance of m / z and deviance of rt based on an objective function
}
\details{
This condition combines different data sources. However, this may 
not be recommended in most cases, when files were aquired under different 
conditions.
}
\examples{
\dontrun{allocatePrecursor2mz(sd01, sd02, kNN = 10, mzCheck = 1, rtCheck = 30, mzVsRTbalance = 10000)}
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

