% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnedMSP-data.R
\docType{data}
\name{binnedMSP}
\alias{binnedMSP}
\title{Example data for \code{MetCirc}: \code{binnedMSP}}
\format{matrix}
\source{
data("idMSMStoMSP.RData", package = "MetCirc")
## create vector with compartments
compSPL <- rep("SPL", length(convert2MSP(tissueSPL)))
compLIM <- rep("LIM", length(convert2MSP(tissueLIM)))
compANT <- rep("ANT", length(convert2MSP(tissueANT)))
compSTY <- rep("STY", length(convert2MSP(tissueSTY)))

compartment <- c(compSPL, compLIM, compANT, compSTY)
binnedMSP <- binning(msp = finalMSP, tol = 0.01, 
                     group = compartment, method = "median")
save(binnedMSP, file = "binnedMSP.RData")
}
\usage{
binnedMSP
}
\value{
matrix
}
\description{
The object \code{binnedMSP} is a matrix, where rows are 
metabolites detected in the tissues sepal (SPL), limb (LIM), anther (ANT) 
and style (STY). The columns contain binned m/z values. Entries 
contain the intensity (in percent) of a certain metabolite at a 
certain m/z value. \code{binnedMSP} is derived from the object \code{tissue} 
and \code{compartmentTissue}.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

