% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{Bin m/z values}
\usage{
binning(msp, tol = 0.01, group)
}
\arguments{
\item{msp}{data.frame in msp format, see ?convert2MSP for further information}

\item{tol}{numerical, boundary value until which neighboured peaks will be 
joined together}

\item{group}{character vector, to which group does the entry belong to}
}
\value{
binning returns a matrix where rownames are precursor ions
(m/z / retention time) and colnames are newly calculated m/z values which 
were binned. The algorithm which is currently implemented joins the 
two nearest m / z values (here the m / z values can also be joined by 
several fragment ions) and recalculates the new m / z by weighing for the 
number of m / z fragments.
}
\description{
Bin m/z values
}
\details{
The functions bins fragments together by calculating
}
\examples{
data("idMSMStoMSP", package = "MetabolomicTools")
group <- sample(c("yl", "ol", "s","r"), size = length(finalMSP), replace=TRUE) 
binning(msp = finalMSP, tol = 0.01, group = group)
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

