% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{classes<-}
\alias{classes<-}
\alias{classes<-,MSP,character-method}
\title{classes<- sets information in MSP object}
\format{An object of class \code{NULL} of length 0.}
\arguments{
\item{x}{object of class MSP, see ?convert2MSP for further information}

\item{value}{character vector with new classes}
}
\value{
MSP
}
\description{
classes<- sets information in MSP object
}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", 
                         splitIndMZ = 2, splitIndRT = NULL)
classes(finalMSP) <- rep("Unknown")
}
\keyword{datasets}

