% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,MSP-method}
\title{combine method for MSP class}
\usage{
\S4method{combine}{MSP,MSP}(object1, object2)
}
\arguments{
\item{object1}{object of class MSP}

\item{object2}{object of class MSP}
}
\value{
MSP object
}
\description{
Combines two objects of class MSP.
}

