% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{compare_Spectra}
\alias{compare_Spectra}
\title{Create similarity matrix from MSnbase::Spectra object}
\usage{
compare_Spectra(object, fun, ...)
}
\arguments{
\item{object}{\code{Spectra}}

\item{fun}{function or character, see ?MSnbase::compareSpectra for further
information}

\item{...}{arguments passed to compareSpectra}
}
\description{
compare_Spectra creates a similarity matrix of all 
Spectrum objects in \code{object}
}
\details{
Function inspired by compareSpectra.OnDiskMSnExp. Possibly
transfer to MSnbase.
}
\examples{
data("spectra", package = "MetCirc")
compare_Spectra(spectra_tissue[1:10], fun = "dotproduct")
}
\author{
Thomas Naake (inspired by compareSpectra.OnDiskMSnExp)
}
