% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\name{convert2MSP}
\alias{convert2MSP}
\title{Convert deconvoluted matrix into MSP format}
\usage{
convert2MSP(mm, splitPattern = "_", splitInd = 1)
}
\arguments{
\item{mm}{matrix, mm has four columns, the first column contains the m/z 
value, the second column the rt, the third column the intensity, the fourth
column the pcgroup_precursorMZ}

\item{splitPattern}{character, splitPattern is the pattern which separates elements
and precursor m/z}

\item{splitInd}{numeric, the position of the precursor m/z concerning separation 
by splitPattern}
}
\description{
Convert deconvoluted matrix into MSP format
}
\details{
Creates a data entry for each precursor ion. Each entry in the 
return object has the following information: NAME, RETENTIONTIME, PRECURSORMZ, 
METABOLITENAME, ADDUCTIONNAME, Num Peaks and a list of fragments together 
with their intensities.
}
\examples{
\dontrun{convert2MSP(mm, splitPattern = "_", splitInd = 1)}
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

