% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertExampleDF-data.R
\docType{data}
\name{convertExampleDF}
\alias{convertExampleDF}
\title{Example data for allocatePrecursor2mz}
\format{data.frame}
\source{
internal
}
\usage{
convertExampleDF
}
\value{
data.frame
}
\description{
This data set contains the object \code{convertExampleDF}.
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

