% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkMatrix.R
\name{createLink0Matrix}
\alias{createLink0Matrix}
\title{Create a link matrix}
\usage{
createLink0Matrix(similarityMatrix, dfNameGroup)
}
\arguments{
\item{similarityMatrix}{matrix, a similarity matrix that contains the 
NDP similarity measure between all precursors in the data set}

\item{dfNameGroup}{data.frame, data.frame contains column "group" and "name"}
}
\value{
createLink0Matrix returns a matrix that gives per each row 
information on linked features
}
\description{
Create a link matrix which links every feature in similarity
matrix with another.
}
\details{
createLink0Matrix creates a matrix from a similarityMatrix which 
includes all connections between features in the similarityMatrix, but 
exclude links which have a similarity of exactly 0.
}
\examples{
data("binnedMSP", package = "MetCirc")
namesPrec <- rownames(binnedMSP)
similarityMat <- createSimilarityMatrix(binnedMSP)
namesPrec <- rownames(binnedMSP)
dfNameGroup <- data.frame(group = unlist(
 lapply(strsplit(namesPrec, "_"), "[[", 1)), name = namesPrec)
dfNameGroup <- dfNameGroup[order(dfNameGroup[,"group"]),] 
createLink0Matrix(similarityMatrix = similarityMat,  
     dfNameGroup = dfNameGroup)
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

