% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkMatrix.R
\name{createLinkMatrix}
\alias{createLinkMatrix}
\title{Create a matrix which contains features to link (indices)}
\usage{
createLinkMatrix(similarityMatrix, threshold, dfNameGroup)
}
\arguments{
\item{similarityMatrix}{matrix, a similarity matrix that contains the 
NDP similarity measure between all precursors in the data set}

\item{threshold}{numerical, threshold value for NDP values, below this value 
linked features will not be included}

\item{dfNameGroup}{data.frame, data.frame contains column "group" and "name"}
}
\description{
Create a matrix which contains features to link (indices)
}
\examples{
\dontrun{createLinkMatrix(similarityMatrix, threshold, dfNameGroup)}
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

