% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{createOrderedSimMat}
\alias{createOrderedSimMat}
\title{Update a similarity matrix according to order of name column in 
dfNameGroup}
\usage{
createOrderedSimMat(dfNameGroup, similarityMatrix)
}
\arguments{
\item{dfNameGroup}{data.frame which contains columns "group" and "name"}

\item{similarityMatrix}{matrix, similarityMatrix contains pair-wise 
similarity coefficients which give information about the similarity between
precursors}
}
\value{
createOrderedSimMat returns a similarity matrix with ordered
rownames according to the name column of dfNameGroup
}
\description{
Internal function for shiny application. May also be used 
outside of shiny to reconstruct figures.
}
\details{
createOrderSimMat takes a dfNameGroup data.frame and a 
similarity matrix as arguments. It will then reorder rows and columns of 
the similarityMatrix object such, that it matches the order of the 
column name in dfNameGroup. createOrderSimMat is used in the shinyCircos 
function to create similarityMatrix objects which will allow to switch
between different types of ordering in between groups (sectors) in the 
circos plot. It may be used as well externally, to reproduce plots outside
of the reactive environment (see vignette for a workflow).
}
\examples{
load(system.file("data/binnedMSP.RData", package = "MetabolomicTools"))
load(system.file("data/similarityMat.RData", package = "MetabolomicTools"))
namesPrec <- rownames(binnedMSP)
dfNameGroup <- data.frame(group = unlist(lapply(strsplit(namesPrec, "_"), "[[", 1)), 
 name = namesPrec) 
## order according to compartment
dfNameGroup <- dfNameGroup[order(dfNameGroup[,"group"]),] 
## order according to retention time and create object dfNameGroupRT
dfNameGroupRT <- orderNames(dfNameGroup, NULL, order = "retentionTime")
createOrderedSimMat(dfNameGroupRT, similarityMatrix = similarityMat)
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

