% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDP.R
\name{createSimilarityMatrix}
\alias{createSimilarityMatrix}
\title{Create similarity matrix}
\usage{
createSimilarityMatrix(mm)
}
\arguments{
\item{mm}{matrix, colnames are all fragments which occur in the dataset, 
rownames are m/z / rt values, entries of mm are intensity values 
corresponding to the mass}
}
\value{
createSimilarityMatrix returns a similarity matrix that contains the 
NDP similarity measure between all precursors in the data set
}
\description{
Creates the similarity matrix by calculating the normalised dot 
product (NDP) between precursors
}
\details{
createSimilarityMatrix calls a function to calculate the 
NDP between all precursors in the data set. For further
information on how the NDP is calculated see ?NDP and Li et al. (2015): 
Navigating natural variation in herbivory-induced secondary metabolism in 
coyote tobacco populations using MS/MS structural analysis. PNAS, 
E4147--E4155.
}
\examples{
data("sd02_deconvoluted", package = "MetabolomicTools")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", splitInd = 2)
compartment <- sample(c("yl", "ol", "s","r"), size = length(finalMSP), 
     replace=TRUE) 
binnedMSP <- binning(msp = finalMSP, tol = 0.01, group = compartment)
similarityMat <- createSimilarityMatrix(binnedMSP)
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

