% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkMatrix.R
\name{cutLinkMatrix}
\alias{cutLinkMatrix}
\title{Create a cut LinkMatrix}
\usage{
cutLinkMatrix(LinkMatrix, type = c("all", "inter", "intra"))
}
\arguments{
\item{LinkMatrix}{matrix, that gives per each row 
information on linked features}

\item{type}{character, one of "all", "inter" or "intra"}
}
\value{
cutLinkMatrix returns a matrix that gives per each row 
information on linked features
}
\description{
Create a cut LinkMatrix
}
\details{
This function is used to cut features from LinkMatrix. If 
type = "all", LinkMatrix will not be changed; if type = "inter" the cut
LinkMatrix will only contain entries of links which are between groups and 
not inside groups; contrary to that, if type = "intra" the cut LinkMatrix 
will only contain entries of links which are inside groups and not between 
groups.
}
\examples{
data("binnedMSP", package = "MetabolomicTools")
namesPrec <- rownames(binnedMSP)
similarityMat <- createSimilarityMatrix(binnedMSP)
namesPrec <- rownames(binnedMSP)
dfNameGroup <- data.frame(group = unlist(lapply(strsplit(namesPrec, "_"), 
                             "[[", 1)), name = namesPrec)
dfNameGroup <- dfNameGroup[order(dfNameGroup[,"group"]),] 
linkMatrix <- createLink0Matrix(similarityMatrix = similarityMat,  
     dfNameGroup = dfNameGroup)
thresholdLinkMatrix(linkMatrix = linkMatrix, threshold = 0.5)
linkMat <- createLinkMatrix(similarityMatrix = similarityMat, threshold = 0.5, 
     dfNameGroup = dfNameGroup)
cutLinkMatrix(LinkMatrix = linkMat, type = "all")
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

