% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\name{cutUniquePrecursor}
\alias{cutUniquePrecursor}
\title{Get unique precursor ions}
\usage{
cutUniquePrecursor(precursor, splitPattern = splitPattern, 
     splitInd = splitInd, returnCharacter = TRUE)
}
\arguments{
\item{precursor}{\code{character} where features are separated by 
\code{splitPattern}}

\item{splitPattern}{\code{character}, character vector to use for splitting, 
see \code{?strsplit} for further information}

\item{splitInd}{\code{numeric}, extract precursor mz at position 
\code{splitInd}}

\item{returnCharacter}{\code{logical}, if \code{TRUE} return 
\code{character}, if \code{FALSE} return \code{numeric}}
}
\value{
The function \code{cutUniquePrecursor} returns \code{character} or \code{numeric} as 
 specified by parameters.
}
\description{
Get unique precursor ions
}
\details{
Function for internal usage.
}
\examples{
precursor <- "A_269.0455469_-1"
splitPattern <- "_"
splitInd <- 2
cutUniquePrecursor(precursor, splitPattern = splitPattern, 
     splitInd = splitInd, returnCharacter = TRUE)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

