% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\name{cutUniquePreMZ}
\alias{cutUniquePreMZ}
\title{Get unique precursor ions}
\usage{
cutUniquePreMZ(precursor, splitPattern = splitPattern, 
     splitInd = splitInd, returnCharacter = TRUE)
}
\arguments{
\item{precursor, }{character with splitPattern}

\item{splitPattern}{character, character vector to use for splitting, 
see ?strsplit for further information}

\item{splitInd}{numeric, extract precursor mz at position splitInd}

\item{returnCharacter}{logical, if TRUE return character, if FALSE 
return numeric}
}
\value{
cutUniquePreMZ returns character as specified by parameters
}
\description{
Get unique precursor ions
}
\details{
Internal function.
}
\examples{
\dontrun{cutUniquePreMZ(precursor, splitPattern = splitPattern, 
     splitInd = splitInd, returnCharacter = TRUE)}
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

