% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,MSP,numeric,missing,missing-method}
\title{Extract parts of a MSP object}
\usage{
\S4method{[}{MSP,numeric,missing,missing}(x, i, j = "missing",
  drop = "missing")
}
\arguments{
\item{x}{object of class MSP}

\item{i}{numeric}

\item{j}{missing}

\item{drop}{missing}
}
\value{
MSP object
}
\description{
[ operator acting on an MSP object to extract parts.
}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", splitInd = 2)
finalMSP[1]
}

