% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{getAdduct}
\alias{getAdduct}
\alias{getAdduct,MSP-method}
\title{getAdduct returns adduct ion names of compounds in MSP 
object}
\format{An object of class \code{NULL} of length 0.}
\usage{
getAdduct(object)
}
\arguments{
\item{object}{object of class MSP}
}
\value{
character
}
\description{
getAdduct returns about adduct ion names of compounds in 
MSP object.
}
\section{Functions}{
\itemize{
\item \code{getAdduct}: returns adduct inon names of compounds in MSP objects
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", 
                         splitIndMZ = 2, splitIndRT = NULL)
getAdduct(finalMSP)
}
\keyword{datasets}

