% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{getBegEndIndMSP}
\alias{getBegEndIndMSP}
\title{Get beginning and end indices of each entry in a data.frame in 
msp format}
\usage{
getBegEndIndMSP(msp)
}
\arguments{
\item{msp}{data.frame in msp format, see ?convert2MSP for further information}
}
\value{
getBegEndIndMSP returns a list of length 2 where the first entry
contains the start indices and the second the end indices
}
\description{
Get beginning and end indices of each entry in a data.frame in 
msp format
}
\details{
Internal use to retrieve indices when fragments start and end.
}
\examples{
\dontrun{getBegEndIndMSP(msp)}
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

