% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{getInformation}
\alias{getInformation}
\alias{getInformation,MSP-method}
\title{getInformation returns information of metabolites in MSP object}
\format{An object of class \code{NULL} of length 0.}
\usage{
getInformation(object)
}
\arguments{
\item{object}{object of class MSP, see ?convert2MSP for further information}
}
\value{
character
}
\description{
getInformation returns information in MSP object.
}
\section{Functions}{
\itemize{
\item \code{getInformation}: returns information of metabolites in MSP objects
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", 
                         splitIndMZ = 2, splitIndRT = NULL)
getInformation(finalMSP)
}
\keyword{datasets}

