% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{getLinkDfIndices}
\alias{getLinkDfIndices}
\title{Get indices in linkDf of feature}
\usage{
getLinkDfIndices(groupnameselected, linkDf)
}
\arguments{
\item{groupnameselected}{`character` vector with groupname of selected 
feature, vector containing "group" and "name" to display, that is 
a unique identifier of the features, "group" and "name" have to be separated
by `"_"` where "group" is the first and "name" is the last element}

\item{linkDf}{`data.frame`, in each row there is information about 
features to be connected}
}
\value{
`getLinkDfIndices` returns indices concerning `linkDf` to which 
`groupnameselected` connects
}
\description{
Gets indices in linkDf of feature
}
\details{
Internal use for function `highlight`
}
\examples{
\dontrun{getLinkDfIndices(groupnameselected, linkMatrix)}

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
