% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{getMetaboliteClass}
\alias{getMetaboliteClass}
\alias{getMetaboliteClass,MSP-method}
\title{getMetaboliteClass returns names of compounds in MSP object}
\format{An object of class \code{NULL} of length 0.}
\usage{
getMetaboliteClass(object)
}
\arguments{
\item{object}{object of class MSP}
}
\value{
character
}
\description{
getMetaboliteClass returns names of compounds in MSP object.
}
\section{Functions}{
\itemize{
\item \code{getMetaboliteClass}: returns class names of compounds in MSP 
objects
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", 
                         splitIndMZ = 2, splitIndRT = NULL)
getMetaboliteClass(finalMSP)
}
\keyword{datasets}

