% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{getMetaboliteName}
\alias{getMetaboliteName}
\alias{getMetaboliteName,MSP-method}
\title{getMetaboliteName returns names of metabolites in MSP object}
\format{An object of class \code{NULL} of length 0.}
\usage{
getMetaboliteName(object)
}
\arguments{
\item{object}{object of class MSP, see ?convert2MSP for further information}
}
\value{
character
}
\description{
getMetaboliteName returns names in MSP object.
}
\section{Functions}{
\itemize{
\item \code{getMetaboliteName}: returns names of metabolites in MSP objects
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", splitInd = 2)
getMetaboliteName(finalMSP)
}
\keyword{datasets}

