% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{getName}
\alias{getName}
\alias{getName,MSP-method}
\title{getName returns names in MSP object}
\format{An object of class \code{NULL} of length 0.}
\usage{
getName(object)
}
\arguments{
\item{object}{object of class MSP, see ?convert2MSP for further information}
}
\value{
character
}
\description{
getName returns names in MSP object.
}
\section{Functions}{
\itemize{
\item \code{getName}: returns names in MSP objects
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", 
                         splitIndMZ = 2, splitIndRT = NULL)
getName(finalMSP)
}
\keyword{datasets}

