% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{getNames}
\alias{getNames}
\alias{getNames,MSP-method}
\title{getNames returns names of compounds in MSP object}
\format{An object of class \code{NULL} of length 0.}
\usage{
getNames(object)
}
\arguments{
\item{object}{object of class MSP}
}
\value{
character
}
\description{
getNames returns names of compounds in MSP object.
}
\section{Functions}{
\itemize{
\item \code{getNames}: returns names of compounds in MSP objects
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", splitInd = 2)
getNames(finalMSP)
}
\keyword{datasets}

