% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idMSMStoMSP-data.R
\docType{data}
\name{idMSMSMtoMSP-data}
\alias{finalMSP}
\alias{idMSMSMtoMSP-data}
\title{Example data for MetabolomicTools}
\format{object of class MSP}
\source{
internal
}
\usage{
finalMSP
}
\value{
object of class MSP
}
\description{
This data set contains the object \code{finalMSP}.
}
\author{
T. Naake
}

