% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{information<-}
\alias{information<-}
\alias{information<-,MSP,character-method}
\title{\code{information<-} sets information in \code{MSP}-object}
\format{An object of class \code{NULL} of length 0.}
\arguments{
\item{x}{object of class \code{MSP}, see \code{?convert2MSP} for further
information}

\item{value}{character vector with new information}
}
\value{
\code{MSP}-object
}
\description{
\code{information<-} sets information in \code{MSP}-object
}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ",
                         splitIndMZ = 2, splitIndRT = NULL)
information(finalMSP) <- rep("Unknown")
}
\keyword{datasets}
