% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{information}
\alias{information}
\alias{information,MSP-method}
\title{information returns information of metabolites in MSP object}
\format{An object of class \code{NULL} of length 0.}
\usage{
information(x)
}
\arguments{
\item{x}{object of class MSP, see ?convert2MSP for further information}
}
\value{
character
}
\description{
information returns information in MSP object.
}
\section{Functions}{
\itemize{
\item \code{information}: returns information of metabolites in MSP object
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", 
                         splitIndMZ = 2, splitIndRT = NULL)
information(finalMSP)
}
\keyword{datasets}

