% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{minFragCart2Polar}
\alias{minFragCart2Polar}
\title{Calculate the nearest feature in polar coordinates given cartesian
coordinates}
\usage{
minFragCart2Polar(x, y, degreeOfFeatures)
}
\arguments{
\item{x}{cartesian x coordinate}

\item{y}{cartesian y coordinate}

\item{degreeOfFeatures}{list of positions of features}
}
\description{
Calculates the nearest feature in polar coordinates given 
 cartesian coordinates
}
\details{
\code{minFragCart2Polar} is employed to find the feature with 
 the smallest distance from given cartesian coordinates.
}
\examples{
\dontrun{minFragCart2Polar(x, y, degreeOfFeatures)}
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

