% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{class}
\name{MSP}
\alias{MSP}
\alias{MSP-class}
\title{MSP-class}
\description{
Definiton of \code{MSP}-class in \code{MetCirc}. Entries are 
MS/MS features including their spectra. 
Allows easy computation of number of entries by entering length(msp), 
where msp is of class MSP. The \code{MSP}-class incorporates accessors 
for auxiliary information of MS/MS features (names, classes, information, 
adduct ion name).
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
