% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\name{msp2FunctionalLossesMSP}
\alias{msp2FunctionalLossesMSP}
\title{Convert MSP to MSP with functional losses}
\usage{
msp2FunctionalLossesMSP(msp)
}
\arguments{
\item{msp}{\code{MSP}-object}
}
\value{
\code{msp2FunctionalLossesMSP} returns a \code{MSP-object}  
(with neutral losses)
}
\description{
\code{msp2FunctionalLossesMSP} converts a \code{MSP}-object 
(with fragments) into a \code{MSP}-object with neutral losses
}
\details{
The function \code{msp2FunctionalLosses} can be used when calculating
the similarity based on neutral losses instead of fragments.
}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", 
                     splitIndMZ = 2, splitIndRT = NULL)
finalMSPNL <- msp2FunctionalLossesMSP(msp = finalMSP)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
