% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{names<-}
\alias{names<-}
\alias{names<-,MSP,character-method}
\title{names<- sets names in MSP object}
\usage{
\S4method{names}{MSP,character}(x) <- value
}
\arguments{
\item{x}{object of class MSP, see ?convert2MSP for further information}

\item{value}{character vector with new names}
}
\value{
MSP
}
\description{
names<- sets names in MSP object
}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", 
                         splitIndMZ = 2, splitIndRT = NULL)
names(finalMSP) <- rep("Unknown")
}

