% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{names}
\alias{names}
\alias{names,MSP-method}
\title{names returns names in MSP object}
\usage{
\S4method{names}{MSP}(x)
}
\arguments{
\item{x}{object of class MSP, see ?convert2MSP for further information}
}
\value{
character
}
\description{
names returns names in MSP object.
}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", 
                         splitIndMZ = 2, splitIndRT = NULL)
names(finalMSP)
}

