% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{orderNames}
\alias{orderNames}
\title{Reorder names in dfNameGroup according to retention time, m/z or 
cluster analysis}
\usage{
orderNames(dfNameGroup, similarityMatrix = NULL, 
     order = c("retentionTime", "mz", "clustering"))
}
\arguments{
\item{dfNameGroup}{data.frame containing column "group" and "name", which is 
a unique identifier of the feature}

\item{similarityMatrix}{matrix, a similarity matrix that contains the 
NDP similarity measure between all precursors in the data set}

\item{order}{character, has to be one of "retentionTime", "mz" or 
"clustering"}
}
\value{
dfNameGroup object (data.frame) with updated (ordered) column 
name
}
\description{
orderNames will order the entries of the column name of the 
dfNameGroup object according to a given sorting instruction.
}
\details{
orderNames takes dfNameGroup, similarityMatrix (only needed
when order = "clustering") and order as arguments. It will rearrange the 
order in dfNameGroup according to the retention time, m/z or 
cluster analysis (based on similarity between fragmentation given by 
entries in the similarityMatrix) and return a reordered dfNameGroup object.
It will manipulate the entries in the name column of dfNameGroup by 
adding consecutively numbered names according to the given order argument. 
orderNames will extract the retention time and mz from the entries of the 
column of dfNameGroup as these entries are composed via the pattern
group_mz/rt. orderNames is used in the shinyCircos function, but can also 
be used externally to reconstruct circos plots outside the reactive 
environment (see vignette for further details).
}
\examples{
load(system.file("data/binnedMSP.RData", package = "MetabolomicTools"))
namesPrec <- rownames(binnedMSP)
dfNameGroup <- data.frame(group = unlist(lapply(strsplit(namesPrec, "_"), "[[", 1)), 
     name = namesPrec) 
## order according to compartment
dfNameGroup <- dfNameGroup[order(dfNameGroup[,"group"]),] 
## order according to retention time
dfNameGroupRT <- orderNames(dfNameGroup, similarityMat, 
 order = "retentionTime")
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

