% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{peaks}
\alias{peaks}
\alias{peaks,MSP-method}
\title{peaks method for MSP class}
\usage{
peaks(object)

\S4method{peaks}{MSP}(object)
}
\arguments{
\item{object}{object of class MSP}
}
\value{
data.frame
}
\description{
Returns the data.frame entry with peak information of an MSP 
object.
}
\section{Methods (by class)}{
\itemize{
\item \code{MSP}: returns the data.frame of an MSP object
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", 
                     splitIndMZ = 2, splitIndRT = NULL)
peaks(finalMSP)
}

