% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotCircos}
\alias{plotCircos}
\title{Circular plot to visualise similarity}
\usage{
plotCircos(dfNameGroup, linkMat, initialize = c(TRUE, FALSE), 
 featureNames = c(TRUE, FALSE), cexFeatureNames = 0.2, 
 groupName = c(TRUE, FALSE), links = c(TRUE, FALSE), highlight = c(TRUE, FALSE))
}
\arguments{
\item{dfNameGroup}{data.frame containing column "group" and "name", which is 
a unique identifier of the feature}

\item{linkMat}{data.frame containing linked features in each row, has five columns 
(group1, name1, group2, name2, NDP)}

\item{initialize}{logical, should plot be initialized?}

\item{featureNames}{logical, should feature names be displayed?}

\item{cexFeatureNames}{numerical, size of feature names}

\item{groupName}{logical, should group names (e.g. compartment names or individual names) be displayed?}

\item{links}{logical, should links be plotted?}

\item{highlight}{logical, are we in highlighting mode?}
}
\description{
Circular plot to visualise similarity
}
\details{
Internal use for shiny app
}
\examples{
\dontrun{plotCircos(dfNameGroup, linkMat, initialize = TRUE, featureNames = TRUE, 
cexFeatureNames = 0.2, groupName = TRUE, links = TRUE)}
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

