% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{printInformationSelect}
\alias{printInformationSelect}
\title{Display information on connected features of selected features}
\usage{
printInformationSelect(groupname, msp = NULL, ind, 
 lMatInd, linkMatrixThreshold, similarityMatrix, roundDigits = 2)
}
\arguments{
\item{groupname}{vector with groupname of selected feature,
vector containing "group" and "name" to display, that is 
a unique identifier of the features, "group" and "name" have to be separated
by "_" where "group" is the first and "name" is the last element}

\item{msp}{MSP, an S4 object of class 'MSP' for information about 
the selected feature}

\item{ind}{numeric}

\item{lMatInd}{numeric indices of selected features}

\item{linkMatrixThreshold}{matrix that contains information of linked 
features of a threshold or greater}

\item{similarityMatrix}{matrix that is used to get information on the degree 
of similarity, similarityMat is an ordered version of a similarity matrix}

\item{roundDigits}{numeric,  how many digits should be displayed?}
}
\value{
character that is in HTML format
}
\description{
Displays information on connected features of selected features.
}
\details{
printInformationSelect is for internal use.
}
\examples{
data("idMSMStoMSP", package = "MetCirc")
data("binnedMSP", package = "MetCirc")
## use only a selection
binnedMSP <- binnedMSP[c(1:20, 29:48, 113:132, 240:259),]
similarityMat <- createSimilarityMatrix(binnedMSP)
groupname <- rownames(similarityMat)
## order similarityMat according to mz
simMat <- createOrderedSimMat(similarityMat, order = "mz") 
groupnameMZ <- rownames(simMat)
linkMat_thr <- createLinkMatrix(simMat, 0.8, 1) 
ind <- 2
indMZ <- which(groupname[ind] == truncateName(groupnameMZ, NULL, group = TRUE))
linkMatInds <- getLinkMatrixIndices(groupnameMZ[indMZ], linkMat_thr)
MetCirc:::printInformationSelect(groupname = groupname, 
 msp = NULL, ind = ind, lMatInd = linkMatInds, 
 linkMatrixThreshold = linkMat_thr, 
 similarityMatrix = similarityMat, roundDigits = 2)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

