% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd01_outputXCMS-data.R
\docType{data}
\name{sd01_outputXCMS}
\alias{sd01_outputXCMS}
\title{Example data for MetCirc}
\format{data.frame}
\source{
internal
}
\usage{
sd01_outputXCMS
}
\value{
data.frame
}
\description{
This data set contains the object \code{sd01_outputXCMS}.
}
\author{
T. Naake, \email{naake@stud.uni-heidelberg.de}
}

