% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd02_deconvoluted-data.R
\docType{data}
\name{sd02_deconvoluted}
\alias{sd02_deconvoluted}
\title{Example data for MetabolomicTools}
\format{data.frame}
\source{
internal
}
\usage{
sd02_deconvoluted
}
\value{
data.frame
}
\description{
This data set contains the object \code{convertExampleDF}.
}
\author{
T. Naake
}

