% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{select}
\alias{select}
\title{Select variable based on condition}
\usage{
select(condition, mz, rt, clust)
}
\arguments{
\item{condition}{`character`, either `"mz"`, `"retentionTime"` or `"clustering"`}

\item{mz}{object to return if `condition == "mz"`}

\item{rt}{object to return if `condition == "retentionTime"`}

\item{clust}{object to return if `condition == "clustering"`}
}
\value{
`mz`, `rt` or `clust` depending on condition
}
\description{
`select` returns `mz`, `rt` or `clust` depending on `condition`.
}
\details{
Helper function for `shinyCircos`, `replayPlotOrder` and `replayPlotAdd`.
}
\examples{
mz <- 1
rt <- 2
clust <- 3
MetCirc:::select(condition = "mz", mz = mz, rt = rt, clust = clust)

}
\author{
Thomas Naake \email{thomasnaake@googlemail.com}
}
