% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{setMetaboliteClass}
\alias{setMetaboliteClass}
\alias{setMetaboliteClass,character,MSP-method}
\title{setMetaboliteClass sets class names of compounds in MSP objects}
\format{An object of class \code{NULL} of length 0.}
\usage{
setMetaboliteClass(object, class)
}
\arguments{
\item{object}{object of class MSP}

\item{class}{character, a vector with new class names}
}
\value{
MSP
}
\description{
setMetaboliteClass sets names of class names of compounds in 
MSP objects. To set names pass a vector with class names to the argument 
\code{class}.
}
\section{Functions}{
\itemize{
\item \code{setMetaboliteClass}: sets class names of compounds in MSP objects
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc") 
finalMSP <- convert2MSP(sd02_deconvoluted, split = "_ ", 
                     splitIndMZ = 2, splitIndRT = NULL)
setMetaboliteClass(finalMSP, c(rep("unknown", 359), "class1"))
}
\keyword{datasets}

