% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{setMetaboliteName}
\alias{setMetaboliteName}
\alias{setMetaboliteName,character,MSP-method}
\title{setMetaboliteName sets metabolite names in MSP objects}
\format{An object of class \code{NULL} of length 0.}
\usage{
setMetaboliteName(object, metName)
}
\arguments{
\item{object}{object of class MSP}

\item{metName}{character, a vector with new metabolite names}
}
\value{
MSP
}
\description{
setMetaboliteName sets metabolite names in MSP objects. To set 
metabolite names pass a vector with names to the argument \code{class}.
}
\section{Functions}{
\itemize{
\item \code{setMetaboliteName}: sets metabolite names in MSP objects
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc") 
finalMSP <- convert2MSP(sd02_deconvoluted, split = "_ ", splitInd = 2)
setMetaboliteName(finalMSP, c(rep("unknown", 358), "met1", "met2"))
}
\keyword{datasets}

