% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{setName}
\alias{setName}
\alias{setName,character,MSP-method}
\title{setName sets names in MSP objects}
\format{An object of class \code{NULL} of length 0.}
\usage{
setName(object, name)
}
\arguments{
\item{object}{object of class MSP}

\item{name}{character, a vector with new names}
}
\value{
MSP
}
\description{
setName sets names in MSP objects. To set names pass a vector 
with names to the argument \code{class}.
}
\section{Functions}{
\itemize{
\item \code{setName}: sets names in MSP objects
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc") 
finalMSP <- convert2MSP(sd02_deconvoluted, split = "_ ", 
                         splitIndMZ = 2, splitIndRT = NULL)
setMetaboliteName(finalMSP, c(rep("unknown", 358), "name1", "name2"))
}
\keyword{datasets}

