% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,MSP-method}
\title{show method for MSP class}
\usage{
\S4method{show}{MSP}(object)
}
\arguments{
\item{object}{object of class MSP}
}
\value{
character
}
\description{
Prints information on the MSP class (number of entries).
}

