% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityMat-data.R
\docType{data}
\name{similarityMat}
\alias{similarityMat}
\title{Example data for MetabolomicTools}
\format{matrix}
\source{
internal
}
\usage{
similarityMat
}
\value{
matrix
}
\description{
This data set contains the object \code{similarityMat}.
}
\author{
T. Naake
}

