% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkMatrix.R
\name{thresholdLinkMatrix}
\alias{thresholdLinkMatrix}
\title{Threshold a link matrix}
\usage{
thresholdLinkMatrix(linkMatrix, threshold)
}
\arguments{
\item{linkMatrix}{matrix, a link matrix that gives per each row 
information on linked features}

\item{threshold}{numerical, threshold value for NDP values, below this value 
linked features will not be returned}
}
\value{
thresholdLinkMatrix returns a matrix that gives per each row 
information on linked features which are linked above a certain threshold
}
\description{
Threshold a link matrix
}
\details{
threshold is a numerical value and filters linked precursor ions; 
filtering is currently based on the normalised dot product.
}
\examples{
load(system.file("data/sd02_deconvoluted.RData", 
     package = "MetabolomicTools")) 
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", splitInd = 2)
binnedMSP <- binning(msp = finalMSP, tol = 0.01)
similarityMat <- createSimilarityMatrix(binnedMSP)
namesPrec <- rownames(binnedMSP)
compartment <- sample(c("yl", "ol", "s","r"), size = length(namesPrec), 
     replace=TRUE) 
namesPrec <- paste(compartment, namesPrec, sep="_")
dfNameGroup <- data.frame(group = compartment, name = namesPrec) 
dfNameGroup <- dfNameGroup[order(dfNameGroup[,"name"]),] 
linkMatrix <- createLink0Matrix(similarityMatrix = similarityMat,  
     dfNameGroup = dfNameGroup)
thresholdLinkMatrix(linkMatrix = linkMatrix, threshold = 0.5)
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

