% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idMSMStissueproject-data.R
\docType{data}
\name{tissue}
\alias{tissue}
\title{Example data for tissue subsection}
\format{data.frame}
\source{
internal
}
\usage{
tissue
}
\value{
data.frame
}
\description{
This data set contains the object \code{tissue}.
}
\author{
T. Naake
}

