% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{truncateName}
\alias{truncateName}
\title{Truncate names}
\usage{
truncateName(dfNameGroup, roundDigits = 2, nameGroup = FALSE)
}
\arguments{
\item{dfNameGroup}{data.frame with group and unique idenfier (name)}

\item{roundDigits}{numeric, how many digits should be displayed?}

\item{nameGroup}{logical, if TRUE name contains group name (e.g. "a_123/456")}
}
\description{
A function to truncate names
}
\details{
The column \code{name} of \code{dfNameGroup} is a vector of \code{character} 
     strings of type consisting of retention time and m/z value. It is 
     cumbersome to display such strings. \code{truncateName} truncates 
     these strings by rounding retention time and m/z values by 
     digits given by \code{roundDigits}. \code{truncateName} is an 
     internal function.
}
\examples{
\dontrun{truncateName(dfNameGroup, roundDigits = 2, nameGroup = FALSE)}
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

