% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{adduct}
\alias{adduct}
\alias{adduct,MSP-method}
\alias{adduct}
\title{\code{adduct} returns adduct ion names of compounds in \code{MSP}-object}
\format{An object of class \code{NULL} of length 0.}
\usage{
adduct(x)
}
\arguments{
\item{x}{object of class \code{MSP}}
}
\value{
\code{character}
}
\description{
\code{adduct} returns adduct ion names of compounds in
\code{MSP}-object.
}
\section{Functions}{
\itemize{
\item \code{adduct}: returns adduct ion names of compounds in
\code{MSP}-objects
}}

\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ",
                         splitIndMZ = 2, splitIndRT = NULL)
adduct(finalMSP)
}
\keyword{datasets}
