% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocatePrecursor2mz.R
\name{allocatePrecursor2mz}
\alias{allocatePrecursor2mz}
\title{allocatePrecursor2mz: Join two data sources}
\usage{
allocatePrecursor2mz(sd01, sd02, kNN = 10, mzCheck = 1, rtCheck = 30, 
     mzVsRTbalance = 10000, splitPattern = "_", splitInd = 2)
}
\arguments{
\item{sd01}{is the output of the \code{XCMS} and \code{CAMERA} 
processing and statistical analysis and \code{XCMS} and \code{CAMERA} 
scripts (see Li et al. 2015 and vignette for further information)}

\item{sd02}{\code{data.frame} with idMS/MS deconvoluted spectra with fragment 
ions (m/z, retention time, relative intensity in \%) and the corresponding 
peak correlation group of the precursor ion. \code{sd02}
has to have at least four columns: a column 'mz', 'rt', 'intensity' and 'id'}

\item{kNN}{numerical, number of k-nearest neighbours based on deviation
from m/z (i.e. the k entries with the smallest deviation)}

\item{mzCheck}{numerical, maximum tolerated distance for m/z (strong 
criterion here)}

\item{rtCheck}{numerical, maximum tolerated distance for retention time}

\item{mzVsRTbalance}{numerical, multiplicator for mz value before calculating 
the (euclidean) distance between two peaks, high value means that there is 
a strong weight on the deviation m/z value}

\item{splitPattern}{character, character vector to use for splitting, see
?strsplit for further information}

\item{splitInd}{numeric, extract precursor mz at position splitInd}
}
\value{
allocatePrecursor2mz returns a \code{data.frame} containing average 
     retention time, average mz, metabolite name, adduct ion name, 
     spectrum reference
}
\description{
Allocates precursor ions to candidate m / z values based on 
minimal distance of m / z and deviance of rt based on an objective function
}
\details{
This function combines different data sources. 
\code{convertExampleDF} is a \code{data.frame} which comprises information 
on a specific metabolite per 
row stating the average retention time, average m/z, the name of the 
metabolite, the adduct ion name, the spectrum 
reference file name and additional information (here: TRIO/LVS). 
\code{allocatePrecursor2mz} uses \code{data.frame}s of the kind of 
\code{sd01\_outputXCMS} and \code{sd02\_deconvoluted} to create a 
\code{data.frame} of the kind of \code{convertExampleDF}. Allocation of 
precursor ions to candidate m/z values is based on minimal distance of m/z 
and deviance of retention time based on an objective function. We can specify 
threshold values for m/z and retention time to be used in 
\code{allocatePrecursor2mz}, as well as the number of neighbours based on 
deviation from m/z values. Also, we can specify the weight to base the 
selection on the m/z compared to the retention time (\code{mzVsRTbalance}). 
This might be useful because m/z values might differ less than the retention 
time in \code{sd01\_outputXCMS} and \code{sd02\_deconvoluted}. Please note, 
that it might be problematic to compare \code{sd01\_outputXCMS} and 
\code{sd02\_deconvoluted} and allocate precursor ions therefrom, 
especially when data were acquired under different conditions.
}
\examples{
data("sd01_outputXCMS", package = "MetCirc")
data("sd02_deconvoluted", package = "MetCirc") 
data("convertExampleDF", package = "MetCirc")
allocatePrecursor2mz(sd01 = sd01_outputXCMS, sd02 = sd02_deconvoluted, 
     kNN = 10, mzCheck = 1, rtCheck = 30, mzVsRTbalance = 10000, splitPattern = " _ ", splitInd = 2)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
\references{
Li et al. (2015): Navigating natural variation in 
herbivory-induced secondary metabolism in coyote tobacco populations using 
MS/MS structural analysis. PNAS, 112, E4147--E4155, 10.1073/pnas.1503106112.
}

