% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnedMSP-data.R
\docType{data}
\name{binnedMSP}
\alias{binnedMSP}
\title{Example data for MetabolomicTools}
\format{matrix}
\source{
internal
}
\usage{
binnedMSP
}
\value{
matrix
}
\description{
This data set contains the object \code{binnedMSP}.
}
\author{
T. Naake
}

