% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnedMSP-data.R
\docType{data}
\name{binnedMSP}
\alias{binnedMSP}
\title{Example data for \code{MetCirc}: \code{binnedMSP}}
\format{matrix}
\source{
internal
}
\usage{
binnedMSP
}
\value{
matrix
}
\description{
The object \code{binnedMSP} is a matrix, where rows are 
metabolites detected in the tissues sepal (SPL), limb (LIM), anther (ANT) 
and style (STY). The columns contain binned m/z values. Entries 
contain the intensity (in percent) of a certain metabolite at a 
certain m/z value. \code{binnedMSP} is derived from the object \code{tissue} 
and \code{compartmentTissue}.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

