% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{Bin m/z values}
\usage{
binning(msp, tol = 0.01)
}
\arguments{
\item{msp}{data.frame in msp format, see ?convert2MSP for further information}

\item{tol}{numerical, boundary value until which neighboured peaks will be joined together}
}
\description{
Bin m/z values
}
\details{
The functions bins fragments together by calculating
}
\examples{
\dontrun{binning(msp, tol = 0.01)}
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

