% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{Bin m/z values}
\usage{
binning(msp, tol = 0.01, group = NULL, method = c("median", "mean"), verbose = FALSE)
}
\arguments{
\item{msp}{\code{MSP}-object, see ?convert2MSP for further information}

\item{tol}{\code{numerical}, boundary value until which neighboured peaks will be 
joined together}

\item{group}{\code{character} vector, to which group does the entry belong to}

\item{method}{\code{character} vector, method has to be \code{median} or 
\code{mean}}

\item{verbose}{\code{logical} vector, if set to TRUE information will be printed
if groups were not detected}
}
\value{
\code{binning} returns a \code{matrix} where rownames are precursor ions
(m/z / retention time) and colnames are newly calculated m/z values which 
were binned. Entires are intensity values in %.
}
\description{
Bin m/z values
}
\details{
The functions \code{binning} bins fragments together by obtaining bins via 
calculating either mean or medians of fragments which were put in intervals 
according to the \code{tol} parameter.
}
\examples{
data("idMSMStoMSP", package = "MetCirc")
binning(msp = finalMSP, tol = 0.01, group = NULL, method = "median", verbose = FALSE)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
