% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{Bin m/z values}
\usage{
binning(msp, tol = 0.01, group = NULL, method = c("median", "mean"))
}
\arguments{
\item{msp}{data.frame in msp format, see ?convert2MSP for further information}

\item{tol}{numerical, boundary value until which neighboured peaks will be 
joined together}

\item{group}{character vector, to which group does the entry belong to}

\item{method}{character vector, method has to be "median" or "mean"}
}
\value{
binning returns a matrix where rownames are precursor ions
(m/z / retention time) and colnames are newly calculated m/z values which 
were binned.
}
\description{
Bin m/z values
}
\details{
The functions bins fragments together by obtaining bins via 
calculating either mean or medians of fragments which were put in intervals 
according to the \code{tol} parameter.
}
\examples{
data("idMSMStoMSP", package = "MetCirc")
##group <- sample(c("yl", "ol", "s","r"), size = length(finalMSP), replace=TRUE) 
binning(msp = finalMSP, tol = 0.01, group = NULL, method = "median")
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

