% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{circosLegend}
\alias{circosLegend}
\title{Plot a legend for circos plot}
\usage{
circosLegend(groupname, highlight = c(TRUE, FALSE), colour = NULL)
}
\arguments{
\item{groupname}{vector containing "group" and "name" to display, that is 
a unique identifier of the features, "group" and "name" have to be separated
by "_" where "group" is the first and "name" is the last element}

\item{highlight}{logical, should colours be adjusted to highlight settings?}

\item{colour}{NULL or character, colour defines the colours which are used
for plotting, if NULL default colours are used}
}
\value{
The function will open a new plot and display colours together 
     with labels.
}
\description{
circosLegend plots a legend for circos plot using group names .
}
\details{
Internal use for shiny app or outside of shiny to reproduce 
     figures.
}
\examples{
 ## load binnedMSP
 data("binnedMSP", package = "MetCirc")
 ## use only a selection 
 binnedMSP <- binnedMSP[c(1:20, 29:48, 113:132, 240:259),]
 similarityMat <- createSimilarityMatrix(binnedMSP)  
 groupname <- rownames(similarityMat)
 ## plot legend
 circosLegend(groupname, highlight = TRUE, colour = NULL)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

