% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{classes}
\alias{classes}
\alias{classes,MSP-method}
\title{\code{classes} returns class names of compounds in \code{MSP}-object}
\format{An object of class \code{NULL} of length 0.}
\usage{
classes(x)
}
\arguments{
\item{x}{object of class \code{MSP}}
}
\value{
\code{character}
}
\description{
\code{classes} returns class names of compounds in \code{MSP}-object.
}
\section{Functions}{
\itemize{
\item \code{classes}: returns class names of metabolites in \code{MSP}-object
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ",
                         splitIndMZ = 2, splitIndRT = NULL)
classes(finalMSP)
}
\keyword{datasets}

