% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,MSP,MSP-method}
\alias{combine,MSP-method}
\title{combine method for MSP class}
\usage{
combine(object1, object2)

\S4method{combine}{MSP,MSP}(object1, object2)
}
\arguments{
\item{object1}{object of class MSP}

\item{object2}{object of class MSP}
}
\value{
MSP object
}
\description{
Combines two objects of class MSP.
}
\section{Methods (by class)}{
\itemize{
\item \code{object1 = MSP,object2 = MSP}: combines two MSP objects
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP1 <- convert2MSP(sd02_deconvoluted, split = " _ ", splitInd = 2)
finalMSP2 <- convert2MSP(sd02_deconvoluted, split = " _ ", splitInd = 2)
combine(finalMSP1, finalMSP2)
}

