% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,MSP-method}
\alias{combine,MSP,MSP-method}
\title{\code{combine} method for \code{MSP}-class}
\usage{
combine(object1, object2)

\S4method{combine}{MSP,MSP}(object1, object2)
}
\arguments{
\item{object1}{object of class \code{MSP}}

\item{object2}{object of class \code{MSP}}
}
\value{
\code{MSP}-object
}
\description{
\code{combine} combines two objects of class \code{MSP}.
}
\section{Methods (by class)}{
\itemize{
\item \code{object1 = MSP,object2 = MSP}: combines two object of class \code{MSP}
}}

\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP1 <- convert2MSP(sd02_deconvoluted, split = " _ ",
                         splitIndMZ = 2, splitIndRT = NULL)
finalMSP2 <- convert2MSP(sd02_deconvoluted, split = " _ ",
                         splitIndMZ = 2, splitIndRT = NULL)
combine(finalMSP1, finalMSP2)
}
