% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idMSMStissueproject-data.R
\docType{data}
\name{compartmentTissue}
\alias{compartmentTissue}
\title{Example data for \code{MetCirc}: \code{compartmentTissue}}
\format{data.frame}
\source{
internal
}
\usage{
tissue
}
\value{
data.frame
}
\description{
The data.frame \code{compartmentTissue} is used in the 
subsection 'Preparing the tissue data set for analysis' in the vignette 
of \code{MetCirc}. In \code{compartmentTissue}, information on the 
organ-localisation of each MS/MS spectrum is stored.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

