% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idMSMStissueproject-data.R
\docType{data}
\name{compartmentTissue}
\alias{compartmentTissue}
\title{Example data for `MetCirc`: `compartmentTissue`}
\format{`data.frame`}
\source{
internal
}
\value{
`data.frame`
}
\description{
The `data.frame` `compartmentTissue` is used in the 
subsection 'Preparing the tissue data set for analysis' in the vignette 
of `MetCirc`. In `compartmentTissue`, information on the 
organ-localisation of each MS/MS spectrum is stored.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
