% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertExampleDF-data.R
\docType{data}
\name{convertExampleDF}
\alias{convertExampleDF}
\title{Example data for \code{MetCirc}: convertExampleDF}
\format{data.frame}
\source{
internal
}
\usage{
convertExampleDF
}
\value{
data.frame
}
\description{
\code{convertExampleDF} is a \code{data.frame} which comprises 
information on a specific metabolite per row stating the average retention 
time, average m/z, the name of the metabolite, the adduct ion name and the 
spectrum reference file name. The function \code{allocatePrecursor2mz} uses 
\code{data.frame}s of the kind of \code{sd01\_outputXCMS} and 
\code{sd02\_deconvoluted} to create a \code{data.frame} of the kind of 
\code{convertExampleDF}. Allocation of precursor ions to candidate
m/z values is based on minimal distance of m/z and deviance of retention time 
based on an objective function. See \code{?allocatePrecursor2mz} for further
information.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

