% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMSP2MSP.R
\name{convertMSP2MSP}
\alias{convertMSP2MSP}
\title{Convert msp data frame into MSP format}
\usage{
convertMSP2MSP(msp)
}
\arguments{
\item{msp}{data frame, in msp format, has the row entries "Name:", "Rt=",
"Num Peaks:" and information on fragments and peak areas}
}
\value{
convertMSP2MSP returns an object of class MSP
}
\description{
Convert msp data frame into MSP format
}
\details{
msp data frame from other sources.
}
\examples{
\dontrun{convertMSP2MSP(msp = msp)}
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

