% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMSP2MSP.R
\name{convertMSP2MSP}
\alias{convertMSP2MSP}
\title{Convert MSP data frame into object of \code{MSP}-class}
\usage{
convertMSP2MSP(msp)
}
\arguments{
\item{msp}{\code{data.frame}, see \code{Details} for further information.}
}
\value{
\code{convertMSP2MSP} returns an object of class \code{MSP}.
}
\description{
Convert msp data frame into object of \code{MSP}-class
}
\details{
msp is a data frame of a .MSP file, a typical data file for 
MS/MS libraries. The data frame has two columns and contains in the first
column the entries "NAME:", 
"PRECURSORMZ:" (or "EXACTMASS:"), "Num Peaks:"  and information on fragments and 
peak areas/intensities. It may additionally contain row entries:
\code{convertMSP2MSP} will try to find the row entries "RETENTIONTIME:",
"ADDUCTIONNAME:" (or "PRECURSORTYPE:"), "CLASS:" and "INFORMATION:" and 
extract the respective information in the second column.
}
\examples{
data("convertMSP2MSP", package = "MetCirc")
convertMSP2MSP(msp = msp2msp)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
