% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkMatrix.R
\name{createLinkMatrix}
\alias{createLinkMatrix}
\title{Create a matrix which contains features to link (indices)}
\usage{
createLinkMatrix(similarityMatrix, threshold, dfNameGroup)
}
\arguments{
\item{similarityMatrix}{matrix, a similarity matrix that contains the 
NDP similarity measure between all precursors in the data set}

\item{threshold}{numerical, threshold value for NDP values, below this value 
linked features will not be included}

\item{dfNameGroup}{data.frame, data.frame contains column "group" and "name"}
}
\value{
createLinkMatrix returns a matrix that gives per each row 
information on linked features
}
\description{
Create a matrix which contains features to link (indices)
}
\details{
Threshold is a numerical value and filters linked precursor ions; 
filtering is currently based on the normalised dot product.
}
\examples{
load(system.file("data/sd02_deconvoluted.RData", 
     package = "MetabolomicTools")) 
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", splitInd = 2)
binnedMSP <- binning(msp = finalMSP, tol = 0.01)
similarityMat <- createSimilarityMatrix(binnedMSP)
namesPrec <- rownames(binnedMSP)
compartment <- sample(c("yl", "ol", "s","r"), size = length(namesPrec), 
     replace=TRUE) 
namesPrec <- paste(compartment, namesPrec, sep="_")
dfNameGroup <- data.frame(group = compartment, name = namesPrec) 
dfNameGroup <- dfNameGroup[order(dfNameGroup[,"name"]),] 
createLinkMatrix(similarityMatrix = similarityMat, threshold = 0.5, 
     dfNameGroup = dfNameGroup)
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

