% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkMatrix.R
\name{createLinkMatrix}
\alias{createLinkMatrix}
\title{Create a matrix which contains features to link (indices)}
\usage{
createLinkMatrix(similarityMatrix, threshold_low, threshold_high)
}
\arguments{
\item{similarityMatrix}{matrix, a similarity matrix that contains the 
NDP similarity measure between all precursors in the data set}

\item{threshold_low}{numerical, threshold value for NDP values, below this 
value linked features will not be included}

\item{threshold_high}{numerical, threshold value for NDP values, above this 
value linked features will not be included}
}
\value{
createLinkMatrix returns a matrix that gives per each row 
information on linked features
}
\description{
Create a matrix which contains features to link (indices)
}
\details{
threshold_low and threshold_high are numerical values and truncate 
similar/identical precursor ions; similarity is currently based on the 
normalised dot product.
}
\examples{
data("binnedMSP", package = "MetCirc")
## use only a selection 
binnedMSP <- binnedMSP[c(c(1:20, 29:48, 113:132, 240:259)),]
similarityMat <- createSimilarityMatrix(binnedMSP)
createLinkMatrix(similarityMatrix = similarityMat, 
     threshold_low = 0.5, threshold_high=1)
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

