% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDP.R
\name{createSimilarityMatrix}
\alias{createSimilarityMatrix}
\title{Create similarity matrix}
\usage{
createSimilarityMatrix(mm)
}
\arguments{
\item{mm}{matrix, colnames are all fragments which occur in the dataset, 
rownames are m/z / rt values, entries of mm are intensity values 
corresponding to the mass}
}
\value{
createSimilarityMatrix returns a similarity matrix that contains the 
NDP similarity measure between all precursors in the data set
}
\description{
Creates the similarity matrix by calculating the normalised dot 
product (NDP) between precursors
}
\details{
createSimilarityMatrix calls a function to calculate the 
NDP between all precursors in the data set. For further
information on how the NDP is calculated see ?NDP and Li et al. (2015): 
Navigating natural variation in herbivory-induced secondary metabolism in 
coyote tobacco populations using MS/MS structural analysis. PNAS, 
E4147--E4155.
}
\examples{
data("binnedMSP", package = "MetCirc")
## truncate binnedMSP 
binnedMSP <- binnedMSP[1:28,]
createSimilarityMatrix(binnedMSP)
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

