% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkMatrix.R
\name{cutLinkMatrix}
\alias{cutLinkMatrix}
\title{Create a cut link matrix}
\usage{
cutLinkMatrix(LinkMatrix, type = c("all", "inter", "intra"))
}
\arguments{
\item{LinkMatrix}{\code{matrix}, that gives per each row 
information on linked features}

\item{type}{\code{character}, one of "all", "inter" or "intra"}
}
\value{
cutLinkMatrix returns a matrix that gives per each row 
information on linked features
}
\description{
Create a cut link matrix
}
\details{
This function is used to cut features from LinkMatrix. If 
type = "all", LinkMatrix will not be changed; if type = "inter" the cut
LinkMatrix will only contain entries of links which are between groups and 
not inside groups; contrary to that, if type = "intra" the cut LinkMatrix 
will only contain entries of links which are inside groups and not between 
groups.
}
\examples{
data("binnedMSP", package = "MetCirc")
## use only a selection 
binnedMSP <- binnedMSP[c(c(1:20, 29:48, 113:132, 240:259)),]
similarityMat <- createSimilarityMatrix(binnedMSP)
linkMat <- createLinkMatrix(similarityMatrix = similarityMat, threshold_low = 0.75, threshold_high = 1)
cutLinkMatrix(LinkMatrix = linkMat, type = "all")
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

