% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,MSP,numeric-method}
\title{Extract parts of a \code{MSP}-object}
\usage{
\S4method{[}{MSP,numeric}(x, i)
}
\arguments{
\item{x}{object of class \code{MSP}}

\item{i}{\code{numeric}}
}
\value{
\code{MSP}-object
}
\description{
[ operator acting on an \code{MSP}-object to extract parts.
}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ",
                         splitIndMZ = 2, splitIndRT = NULL)
finalMSP[1]
}

