% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\name{getBegEndIndMSP}
\alias{getBegEndIndMSP}
\title{Get beginning and end indices of each entry in a \code{data.frame} in
\code{peaks(MSP)}-objects}
\usage{
getBegEndIndMSP(msp)
}
\arguments{
\item{msp}{\code{data.frame} in \code{peaks(MSP)}-object, see
\code{?convert2MSP} for further information}
}
\value{
\code{getBegEndIndMSP} returns a list of length 2 where the first
entry contains the start indices and the second the end indices
}
\description{
Get beginning and end indices of each entry in a
\code{data.frame} in a \code{peaks(MSP)}-object
}
\details{
Internal use to retrieve start and end row indices for fragments of
MS/MS features.
}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ",
                         splitIndMZ = 2, splitIndRT = 3)
finalMSPdf <- peaks(finalMSP)
getBegEndIndMSP(finalMSPdf)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

