% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{getLinkDfIndices}
\alias{getLinkDfIndices}
\title{Get indices in linkDf of feature}
\usage{
getLinkDfIndices(groupnameselected, linkDf)
}
\arguments{
\item{groupnameselected}{\code{character} vector with groupname of selected 
feature, vector containing "group" and "name" to display, that is 
a unique identifier of the features, "group" and "name" have to be separated
by \code{"_"} where "group" is the first and "name" is the last element}

\item{linkDf}{\code{data.frame}, in each row there is information about 
features to be connected}
}
\value{
\code{getLinkDfIndices} returns indices concerning 
 \code{linkDf} to which \code{groupnameselected} connects
}
\description{
Gets indices in linkDf of feature
}
\details{
Internal use for function \code{highlight}.
}
\examples{
\dontrun{getLinkDfIndices(groupnameselected, linkMatrix)}
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
