% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{getLinkMatrixIndices}
\alias{getLinkMatrixIndices}
\title{Get indices in LinkMatrix of feature}
\usage{
getLinkMatrixIndices(dfInd, LinkMatrix)
}
\arguments{
\item{dfInd}{row entry of data.frame dfNameGroup (selected feature)}

\item{LinkMatrix}{matrix, in each row there is information about features 
to be connected}
}
\value{
\code{getLinkMatrixIndices} returns indices concerning LinkMatrix to 
     which dfInd connects
}
\description{
Gets indices in LinkMatrix of feature
}
\details{
Internal use for function highlight.
}
\examples{
\dontrun{getLinkMatrixIndices(dfInd, LinkMatrix)}
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

