% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{getLinkMatrixIndices}
\alias{getLinkMatrixIndices}
\title{Get indices in LinkMatrix of feature}
\usage{
getLinkMatrixIndices(groupnameselected, linkMatrix)
}
\arguments{
\item{groupnameselected}{vector with groupname of selected feature,
vector containing "group" and "name" to display, that is 
a unique identifier of the features, "group" and "name" have to be separated
by "_" where "group" is the first and "name" is the last element}

\item{linkMatrix}{matrix, in each row there is information about features 
to be connected}
}
\value{
\code{getLinkMatrixIndices} returns indices concerning linkMatrix to 
     which groupnameselected connects
}
\description{
Gets indices in LinkMatrix of feature
}
\details{
Internal use for function highlight.
}
\examples{
\dontrun{getLinkMatrixIndices(groupnameselected, linkMatrix)}
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

