% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{getPrecursorMZ}
\alias{getPrecursorMZ}
\title{Get precursor m/z values}
\usage{
getPrecursorMZ(msp)
}
\arguments{
\item{msp}{data.frame in msp format, see ?convert2MSP for further information}
}
\value{
getPrecursorMZ returns a character vector with all precursor 
values
}
\description{
Get precursor m/z values of a data.frame in msp format
}
\details{
Internal use to retrieve precursor m/z values.
}
\examples{
\dontrun{getPrecursorMZ(msp)}
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

