% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{getPrecursorMZ}
\alias{getPrecursorMZ}
\alias{getPrecursorMZ,MSP-method}
\title{getPrecursorMZ returns precursor m/z values of an MSP object}
\format{An object of class \code{NULL} of length 0.}
\usage{
getPrecursorMZ(x)
}
\arguments{
\item{x}{object of class MSP}
}
\value{
numeric
}
\description{
getPrecursorMZ returns a numeric vector with 
 precursor m/z values
}
\section{Functions}{
\itemize{
\item \code{getPrecursorMZ}: returns precursor m/z values of an MSP object
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", 
                         splitIndMZ = 2, splitIndRT = NULL)
getPrecursorMZ(finalMSP)
}
\keyword{datasets}

