% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{getRT}
\alias{getRT}
\title{Get precursor RT values}
\usage{
getRT(msp)
}
\arguments{
\item{msp}{data.frame in msp format, see ?convert2MSP for further information}
}
\value{
getRT returns a character vector with all retention time values
}
\description{
Get precursor RT values of a data.frame in msp format
}
\details{
Internal use to retrieve retention time values.
}
\examples{
\dontrun{getRT(msp)}
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

