% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{getRT}
\alias{getRT}
\alias{getRT,MSP-method}
\title{getRT returns precursor RT values of an MSP object}
\format{An object of class \code{NULL} of length 0.}
\usage{
getRT(object)
}
\arguments{
\item{object}{object of class MSP}
}
\value{
numeric
}
\description{
getRT returns a numeric vector with all retention time values
}
\section{Functions}{
\itemize{
\item \code{getRT}: returns precursor RT values of an MSP object
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", 
                     splitIndMZ = 2, splitIndRT = NULL)
getRT(finalMSP)
}
\keyword{datasets}

