% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{highlight}
\alias{highlight}
\title{Add links and highlight sectors}
\usage{
highlight(dfNameGroup, ind, LinkMatrix)
}
\arguments{
\item{dfNameGroup}{data.frame , data.frame with group and unique idenfier (name)}

\item{ind}{numerical, indices which will be highlighted}

\item{LinkMatrix}{matrix, in each row there is information about features to be connected}
}
\description{
A function to add links and highlight sectors to an initialised
     and plotted \code{circlize} plot with one track.
}
\details{
Internal use for shiny app.
}
\examples{
\dontrun{plotCircosName(dfNameGroup, LinkMatrix, initialize = TRUE, featureNames = TRUE, groupName = TRUE, links = FALSE, highlight = FALSE); highlight(dfNameGroup, ind, LinkMatrix)}
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

