% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{highlight}
\alias{highlight}
\title{Add links and highlight sectors}
\usage{
highlight(groupname, ind, LinkMatrix, colour = NULL, transparency = 0.4)
}
\arguments{
\item{groupname}{vector containing "group" and "name" to display, that is 
a unique identifier of the features, "group" and "name" have to be separated
by "_" where "group" is the first and "name" is the last element}

\item{ind}{numerical, indices which will be highlighted}

\item{LinkMatrix}{matrix, in each row there is information about features 
to be connected}

\item{colour}{NULL or character, colour defines the colours which are used
for plotting, if NULL default colours are used}

\item{transparency}{numerical, defines the transparency of the colours}
}
\value{
The function will update an existing plot by highlighting a 
 specified sector and connected links.
}
\description{
A function to add links and highlight sectors to an initialised
     and plotted \code{circlize} plot with one track.
}
\details{
Internal use for shiny app.
}
\examples{
 ## load binnedMSP
 data("binnedMSP", package = "MetCirc")
 ## use only a selection 
 binnedMSP <- binnedMSP[c(1:20, 29:48, 113:132, 240:259),]
 similarityMat <- createSimilarityMatrix(binnedMSP)
 ## order similarityMat according to retentionTime and update rownames
 simM <- createOrderedSimMat(similarityMat, order = "retentionTime")
 ## create link matrix
 linkMat <- createLinkMatrix(similarityMatrix = simM, 
     threshold_low = 0.95, threshold_high = 1)
 ## cut link matrix (here: only display links between groups)
 linkMat_cut <- cutLinkMatrix(linkMat, type = "inter")
 ## set circlize parameters
 circos.par(gap.degree = 0, cell.padding = c(0.0, 0, 0.0, 0), 
         track.margin = c(0.0, 0))
 groupname <- rownames(simM)
 ## here: set selectedFeatures arbitrarily
 indSelected <- c(2,23,42,62)
 selectedFeatures <- groupname[indSelected]
 ## actual plotting
 plotCircos(groupname, linkMat_cut, initialize = TRUE, 
     featureNames = TRUE, cexFeatureNames = 0.2, groupSector = TRUE, 
     groupName = FALSE, links = FALSE, highlight = TRUE)
 ## highlight
 highlight(groupname = groupname, ind = indSelected, LinkMatrix = 
         linkMat_cut, colour = NULL, transparency = 0.4)
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

