% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idMSMStoMSP-data.R
\docType{data}
\name{idMSMStoMSP-data}
\alias{finalMSP}
\alias{idMSMStoMSP-data}
\title{Example data for \code{MetCirc}: \code{finalMSP}}
\format{object of class MSP}
\source{
internal
}
\usage{
finalMSP
}
\value{
object of class MSP
}
\description{
\code{finalMSP} is of instance 'MSP', a container for 
MS/MS data. \code{finalMSP} is derived from the object \code{tissue} 
and \code{compartmentTissue}.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

