% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{information}
\alias{information}
\alias{information,MSP-method}
\title{\code{information} returns information of metabolites in
 \code{MSP}-object}
\format{An object of class \code{NULL} of length 0.}
\usage{
information(x)
}
\arguments{
\item{x}{object of class \code{MSP}, see \code{?convert2MSP} for further
information}
}
\value{
\code{character}
}
\description{
\code{information} returns information in \code{MSP}-object.
}
\section{Functions}{
\itemize{
\item \code{information}: returns information of metabolites in
\code{MSP}-object
}}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ",
                         splitIndMZ = 2, splitIndRT = NULL)
information(finalMSP)
}
\keyword{datasets}

