% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{length}
\alias{length}
\alias{length,MSP-method}
\title{length method for MSP class}
\usage{
\S4method{length}{MSP}(x)
}
\arguments{
\item{x}{object of class MSP}
}
\value{
numerical
}
\description{
Gives the number of entries in the MSP object.
}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", 
                         splitIndMZ = 2, splitIndRT = NULL)
length(finalMSP)
}

