% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{class}
\name{MSP}
\alias{MSP}
\alias{MSP-class}
\title{MSP-class}
\arguments{
\item{msp}{a data.frame in msp format}
}
\description{
MSP class for msp data.frame. Allows easy computation of 
length of entries by entering length(msp), where msp is of class MSP.
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

