% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\name{msp2FunctionalLossesMSP}
\alias{msp2FunctionalLossesMSP}
\title{Convert MSP to MSP with functional losses}
\usage{
msp2FunctionalLossesMSP(msp)
}
\arguments{
\item{msp}{data.frame, a data.frame in msp format (with fragments)}
}
\value{
msp2FunctionalLossesMSP returns a data.frame in msp format 
(with neutral losses).
}
\description{
msp2FunctionalLossesMSP converts a data.frame in msp format 
(with fragments) into a data.frame in msp format (with neutral losses)
}
\details{
msp2FunctionalLosses can be used when you want to calculate 
the similarity based on neutral losses instead of fragments
}
\examples{
\dontrun{msp2FunctionalLossesMSP(msp)}
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ", splitInd = 2)
finalMSPNL <- msp2FunctionalLossesMSP(msp = finalMSP)
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

