% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMSP2MSP-data.R
\docType{data}
\name{msp2msp}
\alias{msp2msp}
\title{Example data for \code{MetCirc}: \code{msp2msp}}
\format{matrix}
\source{
http://prime.psc.riken.jp/Metabolomics_Software/MS-DIAL/, truncated
.MSP file of GNPS MS/MS Negative (contains 22 entries): 
http://prime.psc.riken.jp/Metabolomics_Software/MS-DIAL/MSMS-GNPS-Curated-Neg.msp
}
\usage{
msp2msp
}
\value{
data.frame
}
\description{
\code{convertMSP2MSP} contains the object \code{msp2msp} 
that is a data frame in .MSP format, a typical format for MS/MS library 
building. Each entry consists of the metabolite name (NAME), the precursor 
mz (PRECURSORMZ), the retention 
time (RETENTIONTIME), number of peaks (Num Peaks), together with fragments and their 
intensity values. In the example used in the function \code{convertMSP2MSP} 
the matrix \code{msp2msp} is used to construct an object of class \code{MSP}.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

