% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{neutralloss}
\alias{neutralloss}
\title{Calculate similarity based on neutral losses}
\usage{
neutralloss(x, y, m = 0.5, n = 2, ...)
}
\arguments{
\item{x}{`Spectrum2` object from `MSnbase` containing
intensity and m/z values, first MS/MS spectrum}

\item{y}{`Spectrum2` object from `MSnbase` containing 
intensity and m/z values, second MS/MS spectrum}

\item{m}{`numeric`, exponent to calculate peak intensity-based weights}

\item{n}{`numeric`, exponent to calculate m/z-based weights}

\item{...}{further arguments passed to `MSnbase:::bin_Spectra`}
}
\value{
`neutralloss` returns a numeric similarity coefficient between 0 and 1
}
\description{
Calculate similarity based on neutral losses (NLS)
}
\examples{
data("spectra", package = "MetCirc")
x <- spectra_tissue[[1]]
y <- spectra_tissue[[2]]
neutralloss(x, y, m = 0.5, n = 2, binSize = 0.01)

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
