% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{orderSimilarityMatrix}
\alias{orderSimilarityMatrix}
\title{Order columns and rows of a similarity matrix according to 
m/z, retention time and clustering}
\usage{
orderSimilarityMatrix(similarityMatrix, order = c("retentionTime", "mz", "clustering"))
}
\arguments{
\item{similarityMatrix}{\code{matrix}, \code{similarityMatrix} contains 
pair-wise similarity coefficients which give information about the similarity 
between precursors}

\item{group}{\code{logical}, if TRUE group separated by "_" will be cleaved
from rownames/colnames of similarityMatrix and matched against names of 
spectra, if FALSE rownames/colnames of similarityMatrix are taken as are 
and matched against names of spectra}

\item{order}{\code{character}, one of "retentionTime", "mz" or "clustering"}
}
\value{
\code{orderSimilarityMat}rix returns a similarity matrix 
with ordered rownames according to the \code{character} vector given to order
}
\description{
Internal function for shiny application. May also be used 
outside of shiny to reconstruct figures.
}
\details{
\code{orderSimilarityMatrix} takes  a similarity matrix,
spectra (containing information on m/z and retentionTime and a 
\code{character} vector
as arguments. It will then reorder rows and columns of 
the similarityMatrix object such, that it orders rows and columns of 
similarityMatrix according to m/z, retention time or clustering in 
each group. \code{orderSimilarityMatrix} is employed in the shinyCircos 
function to create \code{similarityMatrix} objects which will allow to switch
between different types of ordering in between groups (sectors) in the 
circos plot. It may be used as well externally, to reproduce plots outside
of the reactive environment (see vignette for a workflow).
}
\examples{
data("spectra", package="MetCirc")
similarityMat <- compare_Spectra(spectra_tissue[1:10], 
    fun=normalizeddotproduct, binSize=0.01)
## order according to retention time 
orderSimilarityMatrix(similarityMatrix=similarityMat, 
    spectra_tissue, type="retentionTime", group=FALSE)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
