% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{peaks}
\alias{peaks}
\alias{peaks,MSP-method}
\alias{peaks,MSP-method}
\title{\code{peaks} method for \code{MSP}-class}
\usage{
peaks(object)

\S4method{peaks}{MSP}(object)
}
\arguments{
\item{object}{object of class \code{MSP}}
}
\value{
\code{data.frame}
}
\description{
\code{peaks} returns the \code{data.frame} entry with peak
 information of an \code{MSP} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{MSP}: returns the \code{data.frame} of an \code{MSP}-object
}}

\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ",
                     splitIndMZ = 2, splitIndRT = NULL)
peaks(finalMSP)
}
