% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotCircos}
\alias{plotCircos}
\title{Circular plot to visualise similarity}
\usage{
plotCircos(dfNameGroup, linkMat, initialize = c(TRUE, FALSE), 
     featureNames = c(TRUE, FALSE), cexFeatureNames = 0.2, 
     groupSector = c(TRUE, FALSE), groupName = c(TRUE, FALSE), 
     links = c(TRUE, FALSE), highlight = c(TRUE, FALSE))
}
\arguments{
\item{dfNameGroup}{data.frame containing column "group" and "name", which is 
a unique identifier of the feature}

\item{linkMat}{data.frame containing linked features in each row, has 
five columns (group1, name1, group2, name2, NDP)}

\item{initialize}{logical, should plot be initialized?}

\item{featureNames}{logical, should feature names be displayed?}

\item{cexFeatureNames}{numerical, size of feature names}

\item{groupSector}{logical, should groups be displayed with background 
colours?}

\item{groupName}{logical, should group names (e.g. compartment names or 
individual names) be displayed?}

\item{links}{logical, should links be plotted?}

\item{highlight}{logical, are we in highlighting mode?}
}
\value{
The function will initialize a circlize plot and/or will plot 
 features of a circlize plot.
}
\description{
Circular plot to visualise similarity
}
\details{
Internal use for shiny app
}
\examples{
## load binnedMSP
data("binnedMSP", package = "MetCirc")
## use only a selection 
binnedMSP <- binnedMSP[c(1:20, 29:48, 113:132, 240:259),]
similarityMat <- createSimilarityMatrix(binnedMSP) 
namesPrec <- rownames(binnedMSP)
dfNameGroup <- data.frame(group = unlist(lapply(strsplit(namesPrec, "_"), "[[", 1)), 
     name = namesPrec) 
 
## order according to compartment
dfNameGroup <- dfNameGroup[order(dfNameGroup[,"group"]),] 
dfNameGroupRT <- orderNames(dfNameGroup = dfNameGroup, 
     similarityMatrix = NULL, order = "retentionTime")
     
## create a new similarity matrix with updated rownames
simM <- createOrderedSimMat(dfNameGroupRT, similarityMat)
## create link matrix
linkMat <- createLinkMatrix(similarityMatrix = simM, threshold=0.8,
     dfNameGroup = dfNameGroupRT)
## cut link matrix (here: only display links between groups)
linkMat_cut <- cutLinkMatrix(linkMat, type = "inter")
## set circlize paramters
circos.par(gap.degree = 0, cell.padding = c(0.0, 0, 0.0, 0), 
         track.margin = c(0.0, 0))
## here set selectedFeatures arbitrarily
selectedFeatures <- as.character(dfNameGroupRT[c(1,21,41,61),2])
 
## actual plotting
plotCircos(dfNameGroupRT, linkMat_cut, initialize = TRUE, 
    featureNames = TRUE, cexFeatureNames = 0.2, groupSector = TRUE, 
     groupName = FALSE, links = FALSE, highlight = FALSE)
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

