% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotCircos}
\alias{plotCircos}
\title{Circular plot to visualise similarity}
\usage{
plotCircos(groupname, linkMat, initialize = c(TRUE, FALSE), 
     featureNames = c(TRUE, FALSE), cexFeatureNames = 0.3, 
     groupSector = c(TRUE, FALSE), groupName = c(TRUE, FALSE), 
     links = c(TRUE, FALSE), highlight = c(TRUE, FALSE), colour = NULL,
     transparency = 0.2)
}
\arguments{
\item{groupname}{\code{character} vector containing "group" and "name" to 
display, that is a unique identifier of the features, "group" and "name" have 
to be separated 
by \code{"_"} where "group" is the first and "name" is the last element}

\item{linkMat}{\code{data.frame} containing linked features in each row, has 
five columns (group1, name1, group2, name2, NDP)}

\item{initialize}{\code{logical}, should plot be initialized?}

\item{featureNames}{\code{logical}, should feature names be displayed?}

\item{cexFeatureNames}{\code{numeric}, size of feature names}

\item{groupSector}{\code{logical}, should groups be displayed with background 
colours?}

\item{groupName}{\code{logical}, should group names (e.g. compartment names or 
individual names) be displayed?}

\item{links}{\code{logical}, should links be plotted?}

\item{highlight}{\code{logical}, highlight is set to \code{TRUE}}

\item{colour}{\code{NULL} or \code{character}, colour defines the colours 
which are used for plotting, if \code{NULL} default colours are used}

\item{transparency}{\code{numeric}, defines the transparency of the colours}
}
\value{
The function will initialize a circlize plot and/or will plot 
 features of a circlize plot.
}
\description{
Circular plot to visualise similarity
}
\details{
Internal use for \code{shinyCircos} or used outside of 
 \code{shinyCircos} to reproduce figure
}
\examples{
## load binnedMSP
data("binnedMSP", package = "MetCirc")
## use only a selection 
binnedMSP <- binnedMSP[c(1:20, 29:48, 113:132, 240:259),]
similarityMat <- createSimilarityMatrix(binnedMSP) 
## order similarityMat according to retentionTime
simM <- createOrderedSimMat(similarityMat, order = "retentionTime")
## create link matrix
linkMat <- createLinkMatrix(similarityMatrix = simM, 
     threshold_low=0.8, threshold_high=1)
## cut link matrix (here: only display links between groups)
linkMat_cut <- cutLinkMatrix(linkMat, type = "inter")
## set circlize paramters
circos.par(gap.degree = 0, cell.padding = c(0.0, 0, 0.0, 0), 
         track.margin = c(0.0, 0))
groupname <- rownames(simM)
## actual plotting
plotCircos(groupname, linkMat_cut, initialize = TRUE, 
    featureNames = TRUE, cexFeatureNames = 0.3, groupSector = TRUE, 
     groupName = FALSE, links = FALSE, highlight = FALSE, colour = NULL, 
     transparency = 0.2)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

