% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotCircos}
\alias{plotCircos}
\title{Circular plot to visualise similarity}
\usage{
plotCircos(groupname, linkDf, initialize = c(TRUE, FALSE),
  featureNames = c(TRUE, FALSE), cexFeatureNames = 0.3,
  groupSector = c(TRUE, FALSE), groupName = c(TRUE, FALSE),
  links = c(TRUE, FALSE), highlight = c(TRUE, FALSE), colour = NULL,
  transparency = 0.2)
}
\arguments{
\item{groupname}{`character` vector containing "group" and "name" to
display, that is a unique identifier of the features, "group" and "name" have
to be separated
by `"_"` where "group" is the first and "name" is the last element}

\item{linkDf}{`data.frame` containing linked features in each row, has
five columns (group1, spectrum1, group2, spectrum2, similarity)}

\item{initialize}{`logical`, should plot be initialized?}

\item{featureNames}{`logical`, should feature names be displayed?}

\item{cexFeatureNames}{`numeric` size of feature names}

\item{groupSector}{`logical`, should groups be displayed with background
colours?}

\item{groupName}{`logical`, should group names (e.g. compartment names or
individual names) be displayed?}

\item{links}{`logical`, should links be plotted?}

\item{highlight}{`logical`, highlight is set to `TRUE`}

\item{colour}{`NULL` or `character`, colour defines the colours
which are used for plotting, if `NULL` default colours are used}

\item{transparency}{`numeric`, defines the transparency of the colours}
}
\value{
The function will initialize a circlize plot and/or will plot
 features of a circlize plot.
}
\description{
Circular plot to visualise similarity.
}
\details{
Internal use for `shinyCircos` or used outside of
`shinyCircos` to reproduce figure
}
\examples{
data("spectra", package = "MetCirc")
similarityMat <- compare_Spectra(spectra_tissue[1:10],
    fun = normalizeddotproduct, binSize = 0.01)
## order similarityMat according to retentionTime
simM <- orderSimilarityMatrix(similarityMat, spectra = spectra_tissue[1:10],
            type = "retentionTime")
## create link data.frame
linkDf <- createLinkDf(similarityMatrix = simM, spectra = spectra_tissue,
     condition = c("SPL", "LIM", "ANT", "STY"), lower = 0.5, upper = 1)
## cut link data.frame (here: only display links between groups)
linkDf_cut <- cutLinkDf(linkDf, type = "inter")
## set circlize paramters
circos.par(gap.degree = 0, cell.padding = c(0.0, 0, 0.0, 0),
         track.margin = c(0.0, 0))
groupname <- c(as.character(linkDf_cut[, "spectrum1"]),
                as.character(linkDf_cut[, "spectrum2"]))
groupname <- unique(groupname)
## actual plotting
plotCircos(groupname, linkDf_cut, initialize = TRUE,
    featureNames = TRUE, cexFeatureNames = 0.3, groupSector = TRUE,
     groupName = FALSE, links = FALSE, highlight = FALSE, colour = NULL,
     transparency = 0.2)

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
