% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotSpectra}
\alias{plotSpectra}
\title{Plot pair-wise spectra}
\usage{
plotSpectra(spectra, subject, query)
}
\arguments{
\item{spectra}{Spectra object}

\item{subject}{character, name of spectra that is aligned against, character
with preceding sample name}

\item{query}{character, name of spectra that is aligned to subject, character
with preceding sample name}
}
\description{
Plot pair-wise spectra
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
