% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotSpectra}
\alias{plotSpectra}
\title{Plot pair-wise spectra}
\usage{
plotSpectra(spectra, subject, query)
}
\arguments{
\item{spectra}{`Spectra`` object}

\item{subject}{character, name of spectra that is aligned against, character
with preceding sample name}

\item{query}{character, name of spectra that is aligned to subject, character
with preceding sample name}
}
\value{
`ggplot2` plot
}
\description{
``
}
\details{
Internally, all intensities are normalized to 100\%.
}
\examples{
data("spectra", package="MetCirc")
plotSpectra(spectra_tissue, subject="SPL_1", query="SPL_2")
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
