% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{printInformationHover}
\alias{printInformationHover}
\title{Display information on connected features of hovered features}
\usage{
printInformationHover(dfNameGroupOrder, msp = NULL, ind, 
 lMatIndHover, linkMatrixThreshold, highlight = c(TRUE, FALSE), 
 similarityMatrix)
}
\arguments{
\item{dfNameGroupOrder}{data.frame, an ordered data.frame derived from 
dfNameGroup which contains columns "group" and "name", the column "name" 
contains entries in the form of group_number_mz/rt}

\item{msp}{MSP, an S4 object of class 'MSP' for information about 
the hovered feature}

\item{ind}{numeric}

\item{lMatIndHover}{numeric indices of connected features}

\item{linkMatrixThreshold}{matrix that contains information of linked 
features of a threshold or greater}

\item{highlight}{logical only return character when set to TRUE}

\item{similarityMatrix}{matrix that is used to get information on the degree 
of similarity, similarityMat is an ordered version of a similarity matrix}
}
\value{
character that is in HTML format
}
\description{
Displays information on connected features of hovered features.
}
\details{
printInformationHover is for internal use.
}
\examples{
data("idMSMStoMSP", package = "MetCirc")
data("binnedMSP", package = "MetCirc")
binnedMSP <- binnedMSP[c(1:20, 29:48, 113:132, 240:259),]
similarityMat <- createSimilarityMatrix(binnedMSP)
namesPrec <- rownames(binnedMSP)
dfNameGroup <- data.frame(group = unlist(lapply(strsplit(namesPrec, "_"),
                                                "[[", 1)), name = namesPrec)
## order according to compartment
dfNameGroup <- dfNameGroup[order(dfNameGroup[,"group"]),]
dfNameGroupOrder <- orderNames(dfNameGroup, order = "mz")
simMatO <- createOrderedSimMat(dfNameGroupOrder, similarityMat)
linkMat_thr <- createLinkMatrix(simMatO, dfNameGroupOrder, 0.8) 
ind <- 2
linkMatIndsHover <- getLinkMatrixIndices(dfNameGroupOrder[ind,], linkMat_thr)
MetCirc:::printInformationHover(dfNameGroupOrder = dfNameGroupOrder, 
 msp = NULL, ind = ind, lMatIndHover = linkMatIndsHover, 
 linkMatrixThreshold = linkMat_thr, highlight = TRUE, 
 similarityMatrix = simMatO)
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

