% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{printInformationHover}
\alias{printInformationHover}
\title{Display information on connected features of hovered features}
\usage{
printInformationHover(groupname, msp = NULL, ind, 
 lMatIndHover, linkMatrixThreshold, similarityMatrix)
}
\arguments{
\item{groupname}{vector with groupname of selected feature,
vector containing "group" and "name" to display, that is 
a unique identifier of the features, "group" and "name" have to be separated
by "_" where "group" is the first and "name" is the last element}

\item{msp}{MSP, an S4 object of class 'MSP' for information about 
the hovered feature}

\item{ind}{numeric}

\item{lMatIndHover}{numeric indices of connected features}

\item{linkMatrixThreshold}{matrix that contains information of linked 
features of a threshold or greater}

\item{similarityMatrix}{matrix that is used to get information on the degree 
of similarity, similarityMat is an ordered version of a similarity matrix}
}
\value{
character that is in HTML format
}
\description{
Displays information on connected features of hovered features.
}
\details{
printInformationHover is for internal use.
}
\examples{
data("idMSMStoMSP", package = "MetCirc")
data("binnedMSP", package = "MetCirc")
## use only a selection
binnedMSP <- binnedMSP[c(1:20, 29:48, 113:132, 240:259),]
similarityMat <- createSimilarityMatrix(binnedMSP)
## order similarityMat according to mz
simMat <- createOrderedSimMat(similarityMat, order = "mz")
groupname <- rownames(simMat)
linkMat_thr <- createLinkMatrix(simMat, 0.9, 1) 
ind <- 19
linkMatIndsHover <- getLinkMatrixIndices(groupname[ind], linkMat_thr)
MetCirc:::printInformationHover(groupname = groupname, 
 msp = NULL, ind = ind, lMatIndHover = linkMatIndsHover, 
 linkMatrixThreshold = linkMat_thr, 
 similarityMatrix = simMat)
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

