% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{printInformationSelect}
\alias{printInformationSelect}
\title{Display information on connected features of selected features}
\usage{
printInformationSelect(select=groupn[ind], spectra=NULL,
    linkDfInd, linkDf, similarityMatrix, roundDigits=2)
}
\arguments{
\item{select}{`character`}

\item{spectra}{`Spectra` object containing spectra that are compared
in `similarityMatrix`}

\item{linkDfInd}{`numeric` indices of selected features}

\item{linkDf}{`data.frame` that contains information of linked 
features for given thresholds}

\item{similarityMatrix}{`matrix` that is used to get information on the 
degree of similarity, `similarityMat` is an ordered version of a 
similarity matrix, see `?createOrderedSimMat`}

\item{roundDigits}{`numeric`,  how many digits should be displayed?}
}
\value{
`character` that is in HTML format
}
\description{
Displays information on connected features of selected features.
}
\details{
`printInformationSelect` is for internal use.
}
\examples{
data("idMSMStoMSP", package = "MetCirc")
data("binnedMSP", package = "MetCirc")
## use only a selection
binnedMSP <- binnedMSP[c(1:20, 29:48, 113:132, 240:259),]
similarityMat <- createSimilarityMatrix(binnedMSP)
groupname <- rownames(similarityMat)
## order similarityMat according to mz
simMat <- createOrderedSimMat(similarityMat, order = "mz") 
groupnameMZ <- rownames(simMat)
linkMat_thr <- createLinkMatrix(simMat, 0.8, 1) 
ind <- 2
indMZ <- which(groupname[ind] == truncateName(groupnameMZ, NULL, group = TRUE))
linkMatInds <- getLinkDfIndices(groupnameMZ[indMZ], linkMat_thr)
MetCirc:::printInformationSelect(groupname = groupname, 
 msp = NULL, ind = ind, lMatInd = linkMatInds, 
 linkMatrixThreshold = linkMat_thr, 
 similarityMatrix = similarityMat, roundDigits = 2)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
