% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{recordPlotFill_degreeFeatures}
\alias{recordPlotFill_degreeFeatures}
\title{Record a plot of filled features and the degree of features}
\usage{
recordPlotFill_degreeFeatures(type_match, ...)
}
\arguments{
\item{type_match}{`character`, ordered vector according to type}

\item{...}{further arguments passed to `plotCircos`}
}
\value{
`list` of length 2, entry `plotFill` is of `recordedplot` and 
entry `degreeFeatures` that is a `list` of vectors of `numeric(2)`
}
\description{
`recordPlotFill_degreeFeatures` records a plot of filled 
features and returns the degree of features
}
\details{
Helper function for `shinyCircos`.
}
\examples{
type_match <- c("a_1", "a_2", "a_3", "b_1", "b_2", "b_3", "c_1", "c_2")
MetCirc:::recordPlotFill_degreeFeatures(type_match)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
