% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{recordPlotHighlight}
\alias{recordPlotHighlight}
\title{Return a `recordedplot` of `plotCircos` plot with `highlight=TRUE`}
\usage{
recordPlotHighlight(type_match, ...)
}
\arguments{
\item{type_match}{`character`, ordered vector according to type}

\item{...}{further arguments passed to `plotCircos`}
}
\value{
`recordedplot`
}
\description{
`recordPlotHighlight` returns a `recordedplot` object of 
`plotCircos` with `highlight=TRUE`
}
\details{
Helper function for `shinyCircos`.
}
\examples{
type_match <- c("a_1", "a_2", "a_3", "b_1", "b_2", "b_3", "c_1", "c_2")
MetCirc:::recordPlotHighlight(type_match)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
