% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd01_outputXCMS-data.R
\docType{data}
\name{sd01_outputXCMS}
\alias{sd01_outputXCMS}
\title{Example data for `MetCirc`: `sd01_outputXCMS`}
\format{`data.frame`}
\source{
Li et al. (2015)
}
\value{
`data.frame`
}
\description{
`sd01_outputXCMS` is the output file from the package
`XCMS` using the data from Li et al. (2015). See Li et al. (2015) 
for further details.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
