% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd02_deconvoluted-data.R
\docType{data}
\name{sd02_deconvoluted}
\alias{sd02_deconvoluted}
\title{Example data for \code{MetCirc}: sd02_deconvoluted}
\format{\code{data.frame}}
\source{
Li et al. (2015)
}
\usage{
sd02_deconvoluted
}
\value{
\code{data.frame}
}
\description{
\code{sd02_deconvoluted} contains MS/MS data from Li et al. 
 (2015). It is a \code{data.frame} which hosts m/z values, retention time, 
 intensity and the respective precursor m/z values. \code{sd02_deconvoluted} 
 originates from Li et al. (2015). See Li et al. (2015) for further 
 information.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
