% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd02_deconvoluted-data.R
\docType{data}
\name{sd02_deconvoluted}
\alias{sd02_deconvoluted}
\title{Example data for \code{MetCirc}: sd02_deconvoluted}
\format{data.frame}
\source{
Li et al. (2015)
}
\usage{
sd02_deconvoluted
}
\value{
data.frame
}
\description{
sd02_deconvoluted contains MS/MS data from Li et al. (2015). It
is a data.frame which hosts m/z values, retention time, intensity and the
respective precursor m/z values. sd02_deconvoluted originates from 
Li et al. (2015). See Li et al. (2015) for further information.
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

