% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{shinyCircos}
\alias{shinyCircos}
\title{Interactive visualisation of similarity and navigation of MS/MS
features}
\usage{
shinyCircos(similarityMatrix, spectra, condition, ...)
}
\arguments{
\item{similarityMatrix}{`matrix`, `similarityMatrix` contains 
pair-wise similarity coefficients which give information about the similarity 
between MS/MS features}

\item{spectra}{an S4 object of class `Spectra`, the 
`Spectra` object will be used to display information about the selected 
feature and will store information of annotation}

\item{condition}{`character` vector, specifies which condtions/samples
are displayed}

\item{...}{further arguments passed to `shinyCircos`, e.g. 
`cexFeatureNames` to pass to `plotCircos` to set font size in 
`plotCircos` of feature names}
}
\value{
`character`, `shinyCircos` returns a `character` vector with the 
permanently selected precursors and an object with the `Spectra`
object containing the annotation.
}
\description{
Visualise the similarity of MS/MS features in a reactive 
 context. See `Details` the vignette for further descriptions on how to use 
`shinyCircos`.
}
\details{
The function is based on the `shiny` and `circlize` package. 
The user can choose interactively thresholds, type of links (between or 
within groups), display information about MS/MS features, permanently select 
MS/MS features and export selected precursors. The `Spectra` object
stores annotation information about the MS/MS features. Names of features 
within the `similarityMatrix` have to be found as entries 
in `Spectra`. `names(Spectra)` are used as identifiers and 
`colnames`/`rownames` from `similarityMatrix` are cleaved 
by the group identifier (separated by "_"). Annotation information is taken
from `spectra` from the columns "names", "information", "classes" and
"adduct" in the slot `elementMetadata` of `spectra`. After exiting 
the application, the annotation will be written to the respective columns
in the slot `elementMetadata`. If one or several of these columns is 
already present in `elementMetadata`, the column(s) will be used as the 
source of annotation information.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
