% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{shinyCircos}
\alias{shinyCircos}
\title{Interactive visualisation of similarity and navigation of MS/MS features}
\usage{
shinyCircos(similarityMatrix, msp = NULL, ...)
}
\arguments{
\item{similarityMatrix}{\code{matrix}, \code{similarityMatrix} contains 
pair-wise similarity coefficients which give information about the similarity 
between MS/MS features}

\item{spectra}{an S4 object of class \code{Spectra}, the 
\code{Spectra} object will be used to display information about the selected 
feature and will store information of annotation}

\item{...}{further arguments passed to \code{shinyCircos}, e.g. 
\code{cexFeatureNames} to pass to \code{plotCircos} to set font size in 
\code{plotCircos} of feature names}
}
\value{
\code{shinyCircos} returns a \code{character} vector with the 
permanently selected precursors and an object with the \code{Spectra}
object containing the annotation.
}
\description{
Visualise the similarity of MS/MS features in a reactive 
 context. See \code{Details} the vignette for further descriptions on how to use 
 \code{shinyCircos}.
}
\details{
The function is based on the \code{shiny} and \code{circlize} 
package. 
The user can choose interactively thresholds, type of links (between or 
within groups), display information about MS/MS features, permanently select 
MS/MS features and export selected precursors. The \code{Spectra} object
stores annotation information about the MS/MS features. Names of features 
within the \code{similarityMatrix} have to be found as entries 
in \code{Spectra}. \code{names(Spectra)} are used as identifiers and 
\code{colnames}/\code{rownames} from \code{similarityMatrix} are cleaved 
by the group identifier, separated by "_"). Annotation information is taken
from \code{spectra} from the columns "names", "information", "classes" and
"adduct" in the slot elementMetadata of \code{spectra}. After exiting 
the application, the annotation will be written to the respective columns
in the slot elementMetadata. If one or several of these columns is 
already present in elementMetadata, the column(s) will be used as the 
source of annotation information.
}
\examples{
data("idMSMStoMSP", package = "MetCirc")
## truncate files
finalMSP <- finalMSP[c(1:20, 29:48, 113:132, 240:259)]
data("binnedMSP", package = "MetCirc")
binnedMSP <- binnedMSP[c(1:20, 29:48, 113:132, 240:259),]
similarityMat <- createSimilarityMatrix(binnedMSP)
\dontrun{shinyCircos(similarityMatrix = similarityMat, msp = finalMSP)}
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
