% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MSP.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,MSP-method}
\title{\code{show} method for \code{MSP}-class}
\usage{
\S4method{show}{MSP}(object)
}
\arguments{
\item{object}{object of class \code{MSP}}
}
\value{
\code{character}
}
\description{
\code{show} prints information on the \code{MSP}-object (number of entries).
}
\examples{
data("sd02_deconvoluted", package = "MetCirc")
finalMSP <- convert2MSP(sd02_deconvoluted, split = " _ ",
                     splitIndMZ = 2, splitIndRT = NULL)
show(finalMSP)
}

