% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityMat-data.R
\docType{data}
\name{similarityMat}
\alias{similarityMat}
\title{Example data for \code{MetCirc}: \code{similarityMat}}
\format{\code{matrix}}
\source{
data("binnedMSP", package = "MetCirc")
similarityMat <- createSimilarityMatrix(binnedMSP)
save(similarityMat, file = "similarityMat.RData", compress = "xz")
}
\usage{
similarityMat
}
\value{
\code{matrix}
}
\description{
\code{similarityMat} is a \code{matrix} containing the pair-wise
similarity scores derived from the \code{idMSMStissueproject} data set. 
See the vignette for a workflow to reproduce the object \code{similarityMat}.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

