% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityMat-data.R
\docType{data}
\name{similarityMat}
\alias{similarityMat}
\title{Example data for \code{MetCirc}: \code{similarityMat}}
\format{matrix}
\source{
internal
}
\usage{
similarityMat
}
\value{
matrix
}
\description{
\code{similarityMat} is a matrix containing the pair-wise
similarity scores derived from the \code{idMSMStissueproject} data set. 
See the vignette for a workflow to reproduce the object \code{similarityMat}.
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

