% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityMat-data.R
\docType{data}
\name{similarityMat}
\alias{similarityMat}
\title{Example data for `MetCirc`: `similarityMat`}
\format{`matrix`}
\source{
data("spectra", package = "MetCirc")
similarityMat <- compare_Spectra(spectra_tissue,
    fun = normalizeddotproduct, binSize = 0.01)
save(similarityMat, file = "similarityMat.RData", compress = "xz")
}
\value{
`matrix`
}
\description{
`similarityMat` is a `matrix` containing the pair-wise
similarity scores derived from the `idMSMStissueproject` data set.
See the vignette for a workflow to reproduce the object `similarityMat`.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
