% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkDf.R
\name{spectraCond}
\alias{spectraCond}
\title{Get MS/MS spectra that are present in condition}
\usage{
spectraCond(spectra, condition)
}
\arguments{
\item{spectra}{\code{Spectra} object of \code{MSnbase} package}

\item{conditon}{\code{character}, vector with conditions found as columns 
in the elementMetadata slot}
}
\value{
named list with character vector as entries that contains the 
names of the MS/MS entries in spectra that are present in the conditon
(tissues, stress conditions, time points, etc.)
}
\description{
spectraCond returns the names of \code{spectra} that are 
present in condition, corresponding to the slot 
\code{elementMetadata@listData}.
}
\details{
Helper function in createLink0Df and shinyCircos.
}
\examples{
data("spectra", package="MetCirc")
spectraCond(spectra_tissue, condition=c("SPL", "LIM", "ANT", "STY"))
}
\author{
Thomas Naake \email{thomasnaake@googlemail.com}
}
