% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkMatrix.R
\name{thresholdLinkMatrix}
\alias{thresholdLinkMatrix}
\title{Threshold a link matrix}
\usage{
thresholdLinkMatrix(linkMatrix, threshold_low, threshold_high)
}
\arguments{
\item{linkMatrix}{matrix, a link matrix that gives per each row 
information on linked features}

\item{threshold_low}{numerical, threshold value for NDP values, below this 
value linked features will not be returned}

\item{threshold_high}{numerical, threshold value for NDP values, above this 
value linked features will not be returned}
}
\value{
thresholdLinkMatrix returns a matrix that gives per each row 
information on linked features which are linked above a certain threshold
}
\description{
Threshold a link matrix
}
\details{
threshold_low and threshold_high are numerical values and truncates 
similar/identical precursor ions; 
similarity is momentarily based on the normalised dot product.
}
\examples{
data("binnedMSP", package = "MetCirc")
## use only a selection 
binnedMSP <- binnedMSP[c(c(1:20, 29:48, 113:132, 240:259)),]
similarityMat <- createSimilarityMatrix(binnedMSP)
linkMatrix <- createLink0Matrix(similarityMatrix = similarityMat)
thresholdLinkMatrix(linkMatrix = linkMatrix, 
     threshold_low = 0.5, threshold_high=1)
}
\author{
Thomas Naake, \email{naake@stud.uni-heidelberg.de}
}

