% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idMSMStissueproject-data.R
\docType{data}
\name{tissue}
\alias{tissue}
\title{Example data for \code{MetCirc}: \code{tissue}}
\format{\code{data.frame}}
\source{
internal
}
\usage{
tissue
}
\value{
\code{data.frame}
}
\description{
The \code{data.frame} \code{tissue} is used in the subsection
'Preparing the tissue data set for analysis' in the vignette of 
\code{MetCirc}. MS/MS data are merged across floral organs 
in this \code{data.frame}.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

