% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idMSMStissueproject-data.R
\docType{data}
\name{tissue}
\alias{tissue}
\title{Example data for `MetCirc`: `tissue`}
\format{
`data.frame`
}
\source{
internal
}
\value{
`data.frame`
}
\description{
The `data.frame` `tissue` is used in the subsection
'Preparing the tissue data set for analysis' in the vignette of
`MetCirc`. MS/MS data are merged across floral organs
in this `data.frame`.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
