% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{truncateName}
\alias{truncateName}
\title{Truncate names}
\usage{
truncateName(groupname, roundDigits = 2)
}
\arguments{
\item{groupname}{vector with group and unique idenfier (name)}

\item{roundDigits}{numeric, how many digits should be displayed?}
}
\value{
\code{truncateName} returns groupname with truncated names 
without group)
}
\description{
A function to truncate names
}
\details{
\code{groupname} is a vector of \code{character} strings consisting 
     of a group, retention time and m/z value, separated by "_". It is 
     cumbersome to display such long strings. \code{truncateName} 
     truncates these strings by rounding retention time and m/z values by 
     digits given by \code{roundDigits}. \code{truncateName} is an 
     internal function.
}
\examples{
     groupname <- "a_100.12345/10.12345"
     truncateName(groupname, roundDigits = 2)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}

