% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{typeMatch_link0}
\alias{typeMatch_link0}
\title{Get typeMatch and link0 data frame}
\usage{
typeMatch_link0(similarityMatrix, spectra, type, condition)
}
\arguments{
\item{similarityMatrix}{`matrix` with pair-wise similarity values}

\item{spectra}{`Spectra` object}

\item{type}{`character`, either `"mz"`, `"retentionTime"`, `"clustering"`}

\item{condition}{`character`, tissue}
}
\value{
`list` of length 2,  entry `link0df` is of `data.frame` and 
entry `type_match` that is a `character` vector
}
\description{
`typeMatch_link0` returns a list with accessors `"link0df"` and
`"type_match"`
}
\details{
Helper function for `shinyCircos`.
}
\examples{
data("spectra", package = "MetCirc")
similarityMat <- compare_Spectra(spectra_tissue[1:10], 
    fun = normalizeddotproduct, binSize = 0.01)
## order according to retention time 
MetCirc:::typeMatch_link0(similarityMatrix = similarityMat, 
    spectra = spectra_tissue, type = "mz", 
    condition = c("SPL", "LIM", "ANT", "STY"))
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
